Crate stl

STL module

The stl module provides formal definition of “Collections” in terms of traits. Based on primitives exposed by traits, the module provides many generic algorithms for types satisfying “Collection” traits.

The module also implements the “Collection” related traits for stdlib data structures. Currently “Collection” traits have been implemented for:

[T;n] (array)

[T] (slice)

Vec<T> (Vec)

Option<T> (Option)

Range<T> (a..b) where T is a signed/unsigned integer type.

RangeInclusive<T> (a..=b) where T is a signed/unsigned integer type.

&str (string slice).

String (Owned String).

Modules

array

array_impl

array_slice

collections

All the collections exposed from library.

iterators

All the iterators exposed from library.

range

slice_impl

string

string_impl

string_slice

string_slice_impl

value_ref

Proxy Reference to temporary value.

vec_impl

Macros

precondition

Structs

ArraySlice

Unsafe mutable slice for array-like data structures.

Slice

A safe view into unsafe subsequence of a collection.

SliceMut

A safe view into unsafe subsequence of a collection.

StringSlice

Unsafe slice for string.

UnsafeSlice

Traits

BidirectionalCollection

A collection which supports backward traversal as well as forward traversal.

BidirectionalCollectionExt

Algorithms for BidirectionalCollection.

Collection

A multi-pass linear sequence of elements.

CollectionExt

Algorithms for Collection.

LazyCollection

A collection whose elements are computed on element access.

LazyCollectionExt

Algorithms for LazyCollection.

MutableCollection

A collection which allows mutable access to its elements.

MutableCollectionExt

Algorithms for MutableCollection.

ParallelCollectionExt

Parallel Algorithms for Collection.

RandomAccessCollection

A collection that supports efficient random access traversal.

RandomAccessCollectionExt

Algorithms for RandomAccessCollection.

Regular

Any SemiRegular type that is cloneable.

ReorderableCollection

A collection which supports internally reordering its element.

ReorderableCollectionExt

Algorithms for ReorderableCollection.

SemiRegular

Any type that is movable, destructable and equality comparable.

UnsafeMutableSubSequence

A low-level mutable contiguous subsequence of a collection.

UnsafeSubSequence

A low-level, unchecked view into contiguous subsequence of a collection.