stl
Module algo

Algorithms module

The algo module provides a collection of STL algorithms. These algorithms explicitly accepts Positions. For more compact algorithm overloads that just accepts ranges, or infix notation (by .) use rng module.

NOTE: In documentation of any function that accepts position with notation like [start, end), if doc talks in term of full range, it should be considered as rng elements from [start, end) position only.

Functions

adjacent_find_if

Returns first position in range such that element at that position and element after that position satisfies binary predicate.

all_of

Returns true if all elements in range satisfies the predicate.

any_of

Returns true if atleast one element in range satisfies the predicate.

copy

Copies elements from src to out of dest.

copy_if

Copies elements from src to out of dest if it satisfies predicate.

copy_n

Copies n elements from start of src to out of dest.

count

Counts elements in rng equals given element.

count_if

Counts elements in rng that satisfies predicate.

equals

Returns true if rng1 [start1, end1) elements are equal to elements of rng2 [start2, end2) and have same length.

equals_by

Returns true if rng1 [start1, end1) elements are equivalent to elements of rng2 [start2, end2) by relationship bi_pred and have same length.

equals_unbounded

Returns true if rng1 [start1, end1) elements are equal to elements of rng2 starting from start2.

equals_unbounded_by

Returns true if rng1 [start1, end1) elements are equivalent to elements of rng2 starting from start2 by relation bi_pred.

fill_by

Fills rng using given generator.

fill_value

Fills rng with given value.

find

Finds position of first element equals given element.

find_if

Finds position of first element satisfying predicate.

find_if_not

Finds position of first element not satisfying predicate.

is_heap

Returns true if given range is a heap.

is_heap_by

Returns true if given range is a heap wrt comparator.

is_heap_until

Finds largest range that from start represents a heap.

is_heap_until_by

Finds largest range that from start represents a heap wrt comparator.

is_partitioned

Returns true if range is partitioned wrt pred, otherwise false.

make_heap

Reorders the range such that resulting range is heap.

make_heap_by

Reorders the range such that resulting range is heap wrt cmp.

max_element

Returns position of maximum element in the range.

max_element_by

Returns position of maximum element in the range by comparator.

merge

Merges 2 sorted ranges into one sorted range.

merge_by

Merges 2 sorted ranges into one sorted range wrt Comparator.

min_element

Returns position of mimimum element in range.

min_element_by

Returns position of mimimum element in range by comparator.

minmax_element

Returns position of minimum element and maximum element in range.

minmax_element_by

Returns position of minimum element and maximum element in range by comparator.

mismatch

Finds first mismatch position between 2 ranges by equality.

mismatch_by

Finds first mismatch position between 2 ranges by given equivalence relation.

mismatch_unbounded

Finds first mismatch position between 2 ranges by equality where second range is considered to be atleast as big as first one.

mismatch_unbounded_by

Finds first mismatch position between 2 ranges by given equivalence relation where second range is considered to be atleast as big as first one.

none_of

Returns true if no element in range satisfies the predicate.

partition

Partitions range based on given predicate.

partition_point

Returns the position of first such element in partitioned range such that predicate is not satisfied.

pop_heap

Swaps element at start position with element before end position and ensures [start, end - 1) is a heap.

pop_heap_by

Swaps element at start position with element before end position and ensures [start, end - 1) is a heap wrt cmp.

push_heap

Inserts last element in specified range into a heap. After insertion, full range would be a heap.

push_heap_by

Inserts last element in specified range into a heap wrt cmp. After insertion, full range would be heap wrt cmp.

remove

Moves all element equals given element to end of range.

remove_copy

Copies elements from src to dest omitting elements equals given element.

remove_copy_if

Copies elements from src to dest omitting elements satisfying given predicate.

remove_if

Moves all element satisfying predicate to end of range.

replace

Replaces elements of range equals old element with new element.

replace_copy

Copies elements from src to dest with replacing the elements equals given element with new element.

replace_copy_if

Copies elements from src to dest with replacing the elements satisfying predicate with new element.

replace_if

Replaces elements of range satisfying predicate with new element.

reverse

Reverses the given range.

reverse_copy

Copies the given range in reverse order to dest.

rotate

Rotates the given range at mid.

rotate_copy

Copies the given range to dest as if it is rotated at mid.

sort_heap

Converts given heap into sorted range.

sort_heap_by

Converts given heap into sorted range wrt cmp.

stable_partition

Partitions range based on given predicate with preserving relative order of elements.

swap_ranges

Swaps elements of 2 ranges.

transform

Copies elements of src to dest with applying unary operation on it.

unique

Moves all except first of adjacent equal elements to end of range.

unique_by

Moves all except first of adjacent equivalent elements by given equivalence relationship to end of range.

unique_copy

Copies all elements from src to dest with copying only first of adjacent equal elements.

unique_copy_by

Copies all elements from src to dest with copying only first of adjacent equivalent elements by given equivalence relationship.

zip_transform

Copies elements of rng1 and rng2 to dest with applying binary operation on it.