stl::algo
Function binary_search_by

pub fn binary_search_by<Range, Compare>(
    rng: &Range,
    start: Range::Position,
    end: Range::Position,
    ele: &Range::Element,
    is_less: Compare,
) -> bool
where
    Range: ForwardRange + ?Sized,
    Compare: Fn(&Range::Element, &Range::Element) -> bool + Clone,

Checks if element equivalent to ele wrt comparator appears within range.

Precondition

[start, end) represents valid positions in rng.

rng at [start, end) should be partitioned wrt expression rng.at(&i) < ele.

Postcondition

Returns true if element equivalent to ele wrt is_less appears within rng at [start, end), otherwise returns false.

Complexity: O(log2(n)) comparisions. If rng is not a RandomAccessRange, then number of position increment is O(n).

Where n is number of elements in [start, end).

Example

use stl::*;

let arr = [2, 1, 4, 8, 7];
assert!(algo::binary_search_by(&arr, arr.start(), arr.end(), &4, |x, y| x < y));