stl::algo
Function minmax_element_by

pub fn minmax_element_by<Range, Compare>(
    rng: &Range,
    start: Range::Position,
    end: Range::Position,
    cmp: Compare,
) -> (Range::Position, Range::Position)
where
    Range: ForwardRange + ?Sized,
    Compare: Fn(&Range::Element, &Range::Element) -> bool,

Returns position of minimum element and maximum element in range by comparator.

Precondition

[start, end) represents valid positions in rng.

cmp follows strict-weak-ordering relationship.

If a comes before b then cmp(a, b) == true otherwise false.

Postcondition

Returns a pair (i, j) where i denotes the position of minimum element and j denotes the position of maximum element in [start, end) by cmp.

If there are multiple equivalent minimum elements, then i denotes position of first one of them.

If there are multiple equivalent maximum elements, then j denotes position of last one of them.

Complexity: O(n). At most max(floor(3/2(n - 1)), 0) comparisions.

Where n is number of elements in [start, end) of rng.

Example

use stl::*;

let arr = [2, 1, 1, 2, 3, 3, 2];
let (i, j) = algo::minmax_element_by(&arr, arr.start(), arr.end(), |x, y| x < y);
assert_eq!(i, 1);
assert_eq!(j, 5);