stl::algo
Function mismatch_unbounded

pub fn mismatch_unbounded<R1, R2>(
    rng1: &R1,
    start1: R1::Position,
    end1: R1::Position,
    rng2: &R2,
    start2: R2::Position,
) -> (R1::Position, R2::Position)
where
    R1: InputRange + ?Sized,
    R2: InputRange + ?Sized,
    R1::Element: PartialEq<R2::Element>,

Finds first mismatch position between 2 ranges by equality where second range is considered to be atleast as big as first one.

Precondition

[start1, end1) defines valid position of rng1

[start2, start2 + n) defines valid positions of rng2 where n in number of elements in [start1, end1)

bi_pred should follow equivalence relationship.

Postcondition

Returns first position in rng1 and rng2 st rng1.at(p1) != rng2.at(p2)

If no mismatches are found and comparision reaches last1 or last2 whichever happens first, returns positions of that point.

Complexity: O(n), maximum n equality comparisions.