stl::algo
Function nth_element_by

pub fn nth_element_by<Range, Compare>(
    rng: &mut Range,
    start: Range::Position,
    nth: Range::Position,
    end: Range::Position,
    is_less: Compare,
)
where
    Range: RandomAccessRange + SemiOutputRange + ?Sized,
    Compare: Fn(&Range::Element, &Range::Element) -> bool,

Rearranges the elements in range st:

element at nth position is changed to whatever element if range is sorted wrt comparator.

resulting range would be patitioned on nth element wrt comparator.

Precondition

[start, nth) reperesents valid position in rng.

[nth, end) represents valid position in rng.

is_less follows strict-weak-ordering relationship.

Postcondition

element at [start, end) in rng is ordered as such, element at nth position would change to whatever element if rng [start, end) was sorted wrt is_less.

Resulting range rng at [start, end) would be partitioned by element at nth_position, i.e., for any position j > nth, is_less(rng.at(&j), rng.at(&nth)) == false.

if nth == end, do nothing.

Stability of equivalent elements is not guaranteed.

Comlexity: O(n) comparisions on average.

Where n == rng.distance(start, end).

Example

use stl::*;
use rng::infix::*;

let mut arr = [4, 1, 5, 1, 2];
let start = arr.start();
let nth = arr.after_n(start, 2);
let end = arr.end();
algo::nth_element_by(&mut arr, start, nth, end, |x, y| x < y);
assert_eq!(arr.at(&nth), &2);
assert!(&arr[0..nth].is_partitioned(|x| *x < 2));
assert!(&arr[nth..].is_partitioned(|x| *x >= 2));