stl::algo
Function partial_sort_copy_by

pub fn partial_sort_copy_by<SrcRange, DestRange, Compare>(
    src: &SrcRange,
    src_start: SrcRange::Position,
    src_end: SrcRange::Position,
    dest: &mut DestRange,
    dest_start: DestRange::Position,
    dest_end: DestRange::Position,
    is_less: Compare,
) -> DestRange::Position
where
    DestRange: RandomAccessRange + OutputRange + ?Sized,
    SrcRange: InputRange<Element = DestRange::Element> + ?Sized,
    SrcRange::Element: Clone,
    Compare: Fn(&SrcRange::Element, &SrcRange::Element) -> bool + Clone,

Stores d top minimum elements in non decreasing order wrt comparator of source range to destination range where d is number of elements in destination range.

Precondition

[src_start, src_end) represents valid positions in src.

[dest_start, dest_end) represents valid positions in dest.

is_less follows strict-weak-ordering relationship.

Postcondition

Stores d top minimum elements in non decreasing order wrt comparator of src range at [src_start, src_end) to dest range [dest_start, dest_end) where d is number of elements in destination range.

Relative order of equivalent elements are NOT preserved.

Returns the position after last copied element in dest.

Complexity: O(n.log2(min(n, d))) comparisions.

Where n = distance(src_start, src_end), d = distance(dest_start, dest_end).

Example

use stl::*;
use rng::infix::*;

let arr = [4, 1, 5, 1, 2];
let mut dest = [0, 0, 0];
let dest_start = dest.start();
let dest_end = dest.end();
let i = algo::partial_sort_copy_by(&arr, arr.start(), arr.end(), &mut dest, dest_start, dest_end, |x, y| x < y);
assert_eq!(i, 3);
assert!(dest.equals(&[1, 1, 2]));