stl::algo
Function partition

pub fn partition<Range, Predicate>(
    rng: &mut Range,
    start: Range::Position,
    end: Range::Position,
    pred: Predicate,
) -> Range::Position
where
    Range: OutputRange + ?Sized,
    Predicate: Fn(&Range::Element) -> bool,

Partitions range based on given predicate.

Precondition

[start, end) represents valid range in rng.

Postcondition

Reorders elements in rng at [start, end) such that all elements satisfying pred precede elements not satisfying pred.

Relative order of the elements is NOT preserved.

Returns position to first element in modified range that doesn’t satisfy pred.

Complexity: O(n). Exactly n applications of pred. Atmost n swaps.

Where n is number of elements in [start, end).

Example

use stl::*;
use rng::infix::*;

let mut arr = [1, 3, 2, 5, 4];
let start = arr.start();
let end = arr.end();
let i = algo::partition(&mut arr, start, end, |x| x % 2 == 1);
assert_eq!(i, 3);
assert!(arr[0..i].all_of(|x| x % 2 == 1));
assert!(arr[i..].all_of(|x| x % 2 == 0));