stl::algo
Function remove_copy_if

pub fn remove_copy_if<SrcRange, DestRange, Pred>(
    src: &SrcRange,
    start: SrcRange::Position,
    end: SrcRange::Position,
    dest: &mut DestRange,
    out: DestRange::Position,
    pred: Pred,
) -> DestRange::Position
where
    SrcRange: InputRange + ?Sized,
    SrcRange::Element: Clone,
    DestRange: OutputRange<Element = SrcRange::Element> + ?Sized,
    Pred: Fn(&SrcRange::Element) -> bool,

Copies elements from src to dest omitting elements satisfying given predicate.

Precondition

[start, end) represents valid positions in src

dest can accomodate copied elements starting from out.

Postcondition

Copies elements from [start, end) of src to starting from out of dest, omitting the elements which satisfies pred.

Returns position of past the last element copied in dest.

Complexity: O(n). Exactly n applications of pred.

Where n is number of elements in [start, end).

Example

use stl::*;
use rng::infix::*;

let src = [1, 2, 3, 4];

let mut dest = [0, 0];
let out = dest.start();
let i = algo::remove_copy_if(&src, src.start(), src.end(), &mut dest, out, |x| x % 2 == 1);
assert_eq!(i, 2);
assert!(dest.equals(&[2, 4]));