stl::algo
Function reverse_copy

pub fn reverse_copy<SrcRange, DestRange>(
    src: &SrcRange,
    start: SrcRange::Position,
    end: SrcRange::Position,
    dest: &mut DestRange,
    out: DestRange::Position,
) -> DestRange::Position
where
    SrcRange: BidirectionalRange + ?Sized,
    DestRange: OutputRange<Element = SrcRange::Element> + ?Sized,
    SrcRange::Element: Clone,

Copies the given range in reverse order to dest.

Precondition

[start, end) represents valid positions src.

dest must be able to accomodate copied elements starting from out.

Postcondition

Copies elements from [start, end) of src in reverse order to dest starting from out.

Returns position in dest after last copied element.

Complexity: O(n). Exactly n assignments.

Where n is number of elements in [start, end) of src.

Example

use stl::*;
use rng::infix::*;

let src = [1, 2, 3];
let mut dest = [0, 0, 0];
let out = dest.start();
let i = algo::reverse_copy(&src, src.start(), src.end(), &mut dest, out);
assert_eq!(i, 3);
assert!(dest.equals(&[3, 2, 1]));