stl

Module collections

All the collections exposed from library.

Modules

reversed

Structs

EmptyCollection

An empty collection.

LazyMappedCollection

A lazy collection whose elements are applying closure on lazily computed values of base.

MappedCollection

A lazy collection whose elements are applying closure on element-ref of base.

ReversedCollection

A collection that presents element in reverse order of base collection.

SingletonCollection

A collection of one element.