stl
Module rng

Ranges module

The rng module provides a collection of range version of STL algorithms as well as some infix version for some subset of algorithms.

If infix version of range algorithm is supported, then it is in following format:

use stl::*;
use rng::infix::*;

let arr = [1, 2, 3];

let cnt = rng::count_if(&arr, |x| x % 2 == 1);
assert_eq!(cnt, 2);
let cnt = arr.count_if(|x| x % 2 == 1);
assert_eq!(cnt, 2);

For all algorithms it is mentioned in doc if infix version is supported or not for that algorithm. Or open infix module to see all supported infix algorithms.

For using infix version of supported algorithms use:

use stl::rng::infix::*;

Modules

infix

Defines extension traits to support infix version of range algorithms.

Functions

adjacent_find_if

Returns first position in range such that element at that position and element after that position satisfies binary predicate.

all_of

Returns true if all elements in range satisfies the predicate.

any_of

Returns true if atleast one element in range satisfies the predicate.

copy

Copies elements from src to dest.

copy_if

Copies elements from src to out of dest if it satisfies predicate.

count

Counts elements in rng equals given element.

count_if

Counts elements in rng that satisfies predicate.

equals

Returns true if rng1 elements are equal to elements of rng2 and have same length.

equals_by

Returns true if rng1 elements are equivalent to elements of rng2 by relationship bi_pred and have same length.

equals_unbounded

Returns true if rng1 elements are equal to elements of rng2 starting from start2.

equals_unbounded_by

Returns true if rng1 elements are equivalent to elements of rng2 starting from start2 by relation bi_pred.

fill_by

Fills rng using given generator.

fill_value

Fills rng with given value.

find

Finds position of first element equals given element.

find_if

Finds position of first element satisfying predicate.

find_if_not

Finds position of first element not satisfying predicate.

is_partitioned

Returns true if range is partitioned wrt pred, otherwise false.

max_element

Returns position of maximum element in the range.

max_element_by

Returns position of maximum element in the range by comparator.

min_element

Returns position of mimimum element in range.

min_element_by

Returns position of mimimum element in range by comparator.

minmax_element

Returns position of minimum element and maximum element in range.

minmax_element_by

Returns position of minimum element and maximum element in range by comparator.

mismatch

Finds first mismatch position between 2 ranges by equality.

mismatch_by

Finds first mismatch position between 2 ranges by given equivalence relation.

none_of

Returns true if no element in range satisfies the predicate.

partition

Partitions range based on given predicate.

remove

Moves all element equals given element to end of range.

remove_copy

Copies elements from src to dest omitting elements equals given element.

remove_copy_if

Copies elements from src to dest omitting elements satisfying given predicate.

remove_if

Moves all element satisfying predicate to end of range.

replace

Replaces elements of range equals old element with new element.

replace_copy

Copies elements from src to dest with replacing the elements equals given element with new element.

replace_copy_if

Copies elements from src to dest with replacing the elements satisfying predicate with new element.

replace_if

Replaces elements of range satisfying predicate with new element.

reverse

Reverses the given range.

reverse_copy

Copies the given range in reverse order to dest.

rotate

Rotates the given range at mid.

rotate_copy

Copies the given range to dest as if it is rotated at mid.

transform

Copies elements of src to dest with applying unary operation on it.

unique

Moves all except first of adjacent equal elements to end of range.

unique_by

Moves all except first of adjacent equivalent elements by given equivalence relationship to end of range.

unique_copy

Copies all elements from src to dest with copying only first of adjacent equal elements.

unique_copy_by

Copies all elements from src to dest with copying only first of adjacent equivalent elements by given equivalence relationship.

zip_transform

Stores result in dest after applying binary operation on zipped elements of given ranges.