stl::rng
Function partial_sort_copy

pub fn partial_sort_copy<SrcRange, DestRange>(
    src: &SrcRange,
    dest: &mut DestRange,
) -> DestRange::Position
where
    DestRange: RandomAccessRange + OutputRange + BoundedRange + ?Sized,
    SrcRange: InputRange<Element = DestRange::Element> + ?Sized,
    SrcRange::Element: Clone + Ord,

Stores d top minimum elements in non decreasing order of source range to destination range where d is number of elements in destination range.

Precondition

[src_start, src_end) represents valid positions in src.

[dest_start, dest_end) represents valid positions in dest.

Postcondition

Stores d top minimum elements in non decreasing order of src range to dest range where d is number of elements in destination range.

Relative order of equivalent elements are NOT preserved.

Returns the position after last copied element in dest.

Complexity: O(n.log2(min(n, d))) comparisions.

Where n = distance(src_start, src_end), d = distance(dest_start, dest_mid).

Infix Supported

NO

Example

use stl::*;
use rng::infix::*;

let arr = [4, 1, 5, 1, 2];
let mut dest = [0, 0, 0, 0];
let i = rng::partial_sort_copy(&arr, &mut dest);
assert_eq!(i, 4);
assert!(dest.equals(&[1, 1, 2, 4]));