stl

Trait BidirectionalCollection

pub trait BidirectionalCollection: Collection
where
    Self::SubSequence: BidirectionalCollection,
{
    // Required method
    fn form_prior(&self, p: &mut Self::Position);

    // Provided methods
    fn form_prior_n(&self, p: &mut Self::Position, n: usize) { ... }
    fn form_prior_n_limited_by(
        &self,
        position: &mut Self::Position,
        n: usize,
        limit: Self::Position,
    ) -> bool { ... }
    fn prior(&self, p: Self::Position) -> Self::Position { ... }
    fn prior_n(&self, p: Self::Position, n: usize) -> Self::Position { ... }
    fn prior_n_limited_by(
        &self,
        p: Self::Position,
        n: usize,
        limit: Self::Position,
    ) -> Option<Self::Position> { ... }
}

A collection which supports backward traversal as well as forward traversal.

Required Methods

fn form_prior(&self, p: &mut Self::Position)

Sets p to position just before p.

Precondition

Position before p should be well defined.

Provided Methods

fn form_prior_n(&self, p: &mut Self::Position, n: usize)

Sets p to n positions before p.

Precondition

There are n valid positions in self before p.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn form_prior_n_limited_by( &self, position: &mut Self::Position, n: usize, limit: Self::Position, ) -> bool

Increments p by n and returns true; unless the distance is beyond limit in which case set p to limit and returns false.

Precondition

limit should be decrement-reachable from position.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn prior(&self, p: Self::Position) -> Self::Position

Returns position immediately before p

Precondition

Position before p should be well defined.

fn prior_n(&self, p: Self::Position, n: usize) -> Self::Position

Returns nth position before p

Precondition

There are n valid positions before p.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn prior_n_limited_by( &self, p: Self::Position, n: usize, limit: Self::Position, ) -> Option<Self::Position>

Returns nth position before p, unless n position is beyond limit.

Precondition

limit should be increment-reachable from p.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.