stl

Trait BidirectionalCollectionExt

pub trait BidirectionalCollectionExt: BidirectionalCollection
where
    Self::Whole: BidirectionalCollection,
{
    // Provided methods
    fn reverse(&mut self)
       where Self: ReorderableCollection,
             Self::Whole: ReorderableCollection { ... }
    fn fold_right<R, F>(&self, init: R, op: F) -> R
       where F: FnMut(&Self::Element, R) -> R { ... }
}

Algorithms for BidirectionalCollection.

Provided Methods

fn reverse(&mut self)
where Self: ReorderableCollection, Self::Whole: ReorderableCollection,

Reverses the order of elements in self.

Complexity:

O(n)

Example

use stl::*;

let mut arr = [1, 2, 3];
arr.reverse();
assert_eq!(arr, [3, 2, 1]);

fn fold_right<R, F>(&self, init: R, op: F) -> R
where F: FnMut(&Self::Element, R) -> R,

Returns the result of combining elements of given collection using given accumulation operation from right to left.

Postcondition

Result is (e0 + ... + (e(n-1) + (e(n) + init))). where e1, e2, …, en are the references to collection elements.

Complexity:

O(count)

Examples

use stl::*;

let arr = [1, 2, 3];
assert_eq!(arr.fold_right(0, |x, y| x + y), 6);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.