stl
Trait BidirectionalRange

pub trait BidirectionalRange: ForwardRange {
    // Required method
    fn before(&self, i: Self::Position) -> Self::Position;

    // Provided method
    fn before_n(&self, i: Self::Position, n: usize) -> Self::Position { ... }
}

Models a bidirectional range, which can be traversed forward as well as backward.

BidirectionalRange is an extension of ForwardRange.

For supporting backward iteration, before(i) function is supported. before(i) function returns the position just before current position.

A before_n convenient function is there to get nth position before current position.

Required Methods

fn before(&self, i: Self::Position) -> Self::Position

Returns position immediately before i

Precondition

i != start()

Provided Methods

fn before_n(&self, i: Self::Position, n: usize) -> Self::Position

Returns nth position before i

Precondition

self has n valid positions before i.

Complexity

n applications of before.

Object Safety

This trait is not object safe.