stl

Trait Collection

pub trait Collection {
    type Position: Regular;
    type Element;
    type ElementRef<'a>: Deref<Target = Self::Element>
       where Self: 'a;
    type Whole: Collection<Position = Self::Position, Element = Self::Element, Whole = Self::Whole>;
    type Iter<'a>: Iterator<Item = Self::ElementRef<'a>>
       where Self: 'a;

    // Required methods
    fn start(&self) -> Self::Position;
    fn end(&self) -> Self::Position;
    fn form_next(&self, position: &mut Self::Position);
    fn at(&self, i: &Self::Position) -> Self::ElementRef<'_>;
    fn slice(
        &self,
        from: Self::Position,
        to: Self::Position,
    ) -> Slice<'_, Self::Whole>;
    fn iter_within(
        &self,
        from: Self::Position,
        to: Self::Position,
    ) -> Self::Iter<'_>;

    // Provided methods
    fn form_next_n(&self, position: &mut Self::Position, n: usize) { ... }
    fn form_next_n_limited_by(
        &self,
        position: &mut Self::Position,
        n: usize,
        limit: Self::Position,
    ) -> bool { ... }
    fn next(&self, position: Self::Position) -> Self::Position { ... }
    fn next_n(&self, position: Self::Position, n: usize) -> Self::Position { ... }
    fn next_n_limited_by(
        &self,
        position: Self::Position,
        n: usize,
        limit: Self::Position,
    ) -> Option<Self::Position> { ... }
    fn distance(&self, from: Self::Position, to: Self::Position) -> usize { ... }
    fn count(&self) -> usize { ... }
    fn underestimated_count(&self) -> usize { ... }
    fn iter(&self) -> Self::Iter<'_> { ... }

}

Models a multi-pass linear sequence of elements.

Representation:

  _ _ _ _ _ _

  ^            ^
  |            |
start   -->   end

Required Associated Types

type Position: Regular

Type of positions in the collection.

type Element

Type of element in the collection.

type ElementRef<'a>: Deref<Target = Self::Element> where Self: 'a

Type that is like &Element. For collections whose elements are in memory, its simply &Element.

type Whole: Collection<Position = Self::Position, Element = Self::Element, Whole = Self::Whole>

Type representing whole collection. i.e., Self == Slice<W> ? W : Self

type Iter<'a>: Iterator<Item = Self::ElementRef<'a>> where Self: 'a

Type of non-owning iterator yielded by collection.

Required Methods

fn start(&self) -> Self::Position

Returns the position of first element in self, or if self is empty then start() == end()

fn end(&self) -> Self::Position

Returns the position just after last element in collection.

fn form_next(&self, position: &mut Self::Position)

Mutates given position to next position in collection.

Precondition

position != end()

fn at(&self, i: &Self::Position) -> Self::ElementRef<'_>

Access element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

fn slice( &self, from: Self::Position, to: Self::Position, ) -> Slice<'_, Self::Whole>

Returns slice of collection in positions [from, to).

Precondition

[from, to) represents valid positions in collection.

fn iter_within( &self, from: Self::Position, to: Self::Position, ) -> Self::Iter<'_>

Returns a iterator to iterate element-ref in [from, to) positions.

Precondition

[from, to) represents valid positions in collection.

Provided Methods

fn form_next_n(&self, position: &mut Self::Position, n: usize)

Increments given position by n.

Precondition

There are n valid positions in self after position.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn form_next_n_limited_by( &self, position: &mut Self::Position, n: usize, limit: Self::Position, ) -> bool

Increments given position by n, or so that it equals the given limit.

Precondition

limit is a valid position in self.

limit should be increment-reachable from position.

Postcondition

Returns true if position has been incremented exactly by n without going beyond limit, otherwise returns false.

When the return value is false, the value position is equal to limit.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn next(&self, position: Self::Position) -> Self::Position

Returns position immediately after position.

Precondition

position != end()

fn next_n(&self, position: Self::Position, n: usize) -> Self::Position

Returns nth position after position.

Precondition

There are n valid positions in self after position.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn next_n_limited_by( &self, position: Self::Position, n: usize, limit: Self::Position, ) -> Option<Self::Position>

Returns nth position after given position, unless n is beyond limit.

Precondition

limit is a valid position in self.

limit should be increment-reachable from position.

Postcondition

Returns nth position after given position, unless n is beyond limit.

Otherwise, returns None.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn distance(&self, from: Self::Position, to: Self::Position) -> usize

Returns number of elements in [from, to).

Precondition

[from, to) represents valid positions in the collection.

Complexity

O(1) if RandomAccessCollection; O(n) otherwise.

fn count(&self) -> usize

Returns number of elements in collection.

Complexity

O(1) if RandomAccessCollection; O(n) otherwise.

fn underestimated_count(&self) -> usize

Returns count less than or equal to number of elements in collection.

Complexity

O(1) if RandomAccessCollection; O(n) otherwise.

fn iter(&self) -> Self::Iter<'_>

Returns a non-owning iterator to iterate over element-ref of self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.