stl

Trait CollectionExt

pub trait CollectionExt: Collection {
    // Provided methods
    fn iter(&self) -> CollectionIterator<'_, Self::Whole>  { ... }
    fn first(&self) -> Option<<Self as Collection>::ElementRef<'_>> { ... }
    fn all(&self) -> Slice<'_, Self::Whole> { ... }
    fn prefix(&self, to: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn suffix(&self, from: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn equals_by<OtherCollection, F>(
        &self,
        other: &OtherCollection,
        bi_pred: F,
    ) -> bool
       where OtherCollection: Collection,
             F: Fn(&Self::Element, &OtherCollection::Element) -> bool { ... }
    fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
       where OtherCollection: Collection<Element = Self::Element>,
             Self::Element: Eq { ... }
    fn find_if<Pred>(&self, pred: Pred) -> Self::Position
       where Pred: Predicate<Self::Element> { ... }
    fn for_each<F>(&self, f: F)
       where F: FnMut(&Self::Element) { ... }
}

Algorithms for Collection.

Provided Methods

fn iter(&self) -> CollectionIterator<'_, Self::Whole>

Returns a non-consuming iterator that iterates over &Self::Element.

fn first(&self) -> Option<<Self as Collection>::ElementRef<'_>>

Returns the first element, or nil if self is empty.

fn all(&self) -> Slice<'_, Self::Whole>

Returns slice of the collection covering full collection.

Precondition

Postcondition

Returns slice of the collection covering full collection.

Complexity: O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.all();
assert!(s.equals(&[1, 2, 3, 4, 5]));

fn prefix(&self, to: Self::Position) -> Slice<'_, Self::Whole>

Returns prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Postcondition

Returns prefix slice of the collection ending at to exclusive.

Complexity: O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let p = arr.prefix(3);
assert!(p.equals(&[1, 2, 3]));

fn suffix(&self, from: Self::Position) -> Slice<'_, Self::Whole>

Returns suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Postcondition

Returns suffix slice of the collection starting from from inclusive.

Complexity: O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.suffix(3);
assert!(s.equals(&[4, 5]));

fn equals_by<OtherCollection, F>( &self, other: &OtherCollection, bi_pred: F, ) -> bool
where OtherCollection: Collection, F: Fn(&Self::Element, &OtherCollection::Element) -> bool,

Returns true if elements of self is equivalent to elements of other by given relation bi_pred.

Precondition

Postcondition

Returns true if elements of self is equivalent to other by given relation bi_pred.

If self and other have different number of elements, then return false.

Complexity: O(min(m, n)) where

m == self.size()

n == other.size()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [2, 3, 4];
assert!(arr1.equals_by(&arr2, |x, y| *y == x + 1));

fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
where OtherCollection: Collection<Element = Self::Element>, Self::Element: Eq,

Returns true if elements of self is equal to elements of other.

Precondition

Postcondition

Returns true if elements of self is equal to elements of other.

If self and other have different number of elements, then return false.

Complexity: O(min(m, n)) where

m == self.size()

n == other.size()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [1, 2, 3];
assert!(arr1.equals(&arr2));

fn find_if<Pred>(&self, pred: Pred) -> Self::Position
where Pred: Predicate<Self::Element>,

Finds position of first element satisfying predicate.

Precondition

Postcondition

Returns position of first element in self satisfying pred.

Returns end position if no such element exists.

Complexity: O(n). Maximum n applications of pred.

where n is number of elements in self.

Example

use stl::*;

let arr = [1, 2, 3];
let i = arr.find_if(|x| *x == 3);
assert_eq!(i, 2);

fn for_each<F>(&self, f: F)
where F: FnMut(&Self::Element),

Applies f to each element of collection.

Precondition

Postcondition

Applies f to each element of collection.

Complexity: O(n). Exactly n applications of f.

Example

use stl::*;

let arr = [1, 2, 3];
let mut sum = 0;
arr.for_each(|x| sum = sum + x);
assert_eq!(sum, 6);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.