stl

Trait DoubleEndedLendingIterator

pub trait DoubleEndedLendingIterator: LendingIterator {
    // Required method
    fn next_back(&mut self) -> Option<Self::Item<'_>>;
}

A lending iterator whose elements can be consumed from other side too.

Required Methods

fn next_back(&mut self) -> Option<Self::Item<'_>>

Returns next back element if any, and consumes that element.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.