stl
Trait ForwardRange

pub trait ForwardRange: InputRange<Position: Regular> {
    // Provided method
    fn distance(&self, from: Self::Position, to: Self::Position) -> usize { ... }
}

Models a multi-pass range.

ForwardRange is an extension of InputRange.

ForwardRange models multi-pass range by allowing Position to be Regular. This allows cloning position object, so that clone can be used for another iteration at any point.

Usually any algorithm, that needs to store iterator usually have minimum ForwardRange requirement.

Supports a convenient distance function, to get distance between 2 positions.

Provided Methods

fn distance(&self, from: Self::Position, to: Self::Position) -> usize

Returns distance between position [from, to).

Precondition

[from, to) represents valid position in range.

Complexity

O(n)