stl
Trait InputRange

pub trait InputRange {
    type Element;
    type Position: SemiRegular;
    type ElementRef<'a>: Deref<Target = Self::Element>
       where Self: 'a;

    // Required methods
    fn start(&self) -> Self::Position;
    fn is_end(&self, i: &Self::Position) -> bool;
    fn after(&self, i: Self::Position) -> Self::Position;
    fn at<'a>(&'a self, i: &Self::Position) -> Self::ElementRef<'a>;

    // Provided method
    fn after_n(&self, i: Self::Position, n: usize) -> Self::Position { ... }
}

Models a single-pass range.

This is most primitive range trait. It specfies requirement for supporting single pass iteration.

type Element: for defining the type of element in range.

type Position: for defining the type of position/index in range.

start() would return the first position

end() would return the last position.

For traversing forward after(i) would return next position after i.

Accessing any element immutably can be done using at(i)

after_n is additional convinient method supported to get to nth position after current position.

Position type is bounded to SemiRegular so that algorithms working on InputRange doesn’t accidently copy any position.

Required Associated Types

type Element

Type of the element contained in self

type Position: SemiRegular

Type of the positions in self

type ElementRef<'a>: Deref<Target = Self::Element> where Self: 'a

Type of reference to element.

Very useful in scenarios when range actually doesn’t contain the elements and proxy reference is necessary.

Required Methods

fn start(&self) -> Self::Position

Returns the position of first element in self, or if self is empty then start() == end()

fn is_end(&self, i: &Self::Position) -> bool

Returns true if i is the end position of range.

fn after(&self, i: Self::Position) -> Self::Position

Returns position immediately after i

Precondition

i != end()

fn at<'a>(&'a self, i: &Self::Position) -> Self::ElementRef<'a>

Access element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

Provided Methods

fn after_n(&self, i: Self::Position, n: usize) -> Self::Position

Returns nth position after i.

Precondition

There are n valid positions in self after i.

Complexity

n applications of after().

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.