stl

Trait LazyCollection

pub trait LazyCollection: Collection
where
    Self::Whole: LazyCollection,
{
    type LazyIter<'a>: Iterator<Item = Self::Element>
       where Self: 'a;

    // Required methods
    fn compute_at(&self, i: &Self::Position) -> Self::Element;
    fn lazy_iter_within(
        &self,
        from: Self::Position,
        to: Self::Position,
    ) -> Self::LazyIter<'_>;

    // Provided method
    fn lazy_iter(&self) -> Self::LazyIter<'_> { ... }
}

Models a collection whose elements are computed on element access.

Required Associated Types

type LazyIter<'a>: Iterator<Item = Self::Element> where Self: 'a

Type of iterator to iterate of lazy values of self.

Required Methods

fn compute_at(&self, i: &Self::Position) -> Self::Element

Computes element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

fn lazy_iter_within( &self, from: Self::Position, to: Self::Position, ) -> Self::LazyIter<'_>

Returns a lazy iterator that iterate lazy element values in [from, to) positions.

Precondition

[from, to) represents valid positions in collection.

Provided Methods

fn lazy_iter(&self) -> Self::LazyIter<'_>

Returns a lazy iterator that iterate over lazy element values.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.