stl

Trait LazyCollectionExt

pub trait LazyCollectionExt: LazyCollection
where
    Self::Whole: LazyCollection,
{
    // Provided methods
    fn lazy_first(&self) -> Option<Self::Element> { ... }
    fn lazy_last(&self) -> Option<Self::Element>
       where Self: BidirectionalCollection,
             Self::Whole: BidirectionalCollection { ... }
    fn lazy_iter(&self) -> LazyCollectionIter<'_, Self::Whole>  { ... }
    fn lazy_for_each<F>(&self, f: F)
       where F: FnMut(Self::Element) { ... }
    fn lazy_map<MapFn, MappedType>(
        self,
        map_fn: MapFn,
    ) -> LazyMappedCollection<Self, MapFn, MappedType>
       where Self: Sized,
             MapFn: Fn(Self::Element) -> MappedType { ... }
    fn lazy_partitioned<F>(
        &self,
        belongs_in_second_half: F,
    ) -> (Vec<Self::Element>, Vec<Self::Element>)
       where F: FnMut(&Self::Element) -> bool { ... }
    fn lazy_fold_left<R, F>(&self, init: R, op: F) -> R
       where F: FnMut(R, Self::Element) -> R { ... }
    fn lazy_fold_right<R, F>(&self, init: R, op: F) -> R
       where F: FnMut(Self::Element, R) -> R,
             Self: BidirectionalCollection,
             Self::Whole: BidirectionalCollection { ... }
}

Algorithms for LazyCollection.

Provided Methods

fn lazy_first(&self) -> Option<Self::Element>

Returns the “lazily computed” first element, or nil if self is empty.

fn lazy_last(&self) -> Option<Self::Element>
where Self: BidirectionalCollection, Self::Whole: BidirectionalCollection,

Returns the “lazily computed” last element, or nil if self is empty.

fn lazy_iter(&self) -> LazyCollectionIter<'_, Self::Whole>

Returns an iterator to iterate over lazyily computed elements in collection.

fn lazy_for_each<F>(&self, f: F)
where F: FnMut(Self::Element),

Applies f to each “lazily computed” element of collection.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = 1..4;
let mut sum = 0;
arr.lazy_for_each(|x| sum = sum + x);
assert_eq!(sum, 6);

fn lazy_map<MapFn, MappedType>( self, map_fn: MapFn, ) -> LazyMappedCollection<Self, MapFn, MappedType>
where Self: Sized, MapFn: Fn(Self::Element) -> MappedType,

Returns a lazy collection projecting elements of mapping the given closure over lazily computed values of self.

Example

use stl::*;

let arr = (1..=5).lazy_map(|x| x * 2);
assert!(arr.equals(&[2, 4, 6, 8, 10]));

fn lazy_partitioned<F>( &self, belongs_in_second_half: F, ) -> (Vec<Self::Element>, Vec<Self::Element>)
where F: FnMut(&Self::Element) -> bool,

Returns two Vec containing the elements of the collection that don’t and do satisfy the given predicate, respectively.

Postcondition

Returns (falseVec, trueVec) where falseVec contains all elements that don’t satisfy predicate and trueVec contains all elements that do satisfy predicate.

Relative ordering of elements is preserved in both Vec.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let arr = 1..=5;
let (evens, odds) = arr.lazy_partitioned(|x| x % 2 == 1);
assert_eq!(evens, [2, 4]);
assert_eq!(odds, [1, 3, 5]);

fn lazy_fold_left<R, F>(&self, init: R, op: F) -> R
where F: FnMut(R, Self::Element) -> R,

Returns the result of combining elements of given collection using given accumulation operation from left to right.

Postcondition

Result is (((init + e1) + e2) + ... + en). where e1, e2, …, en are the lazily computed values of collection, where (a + b) represents op(a, b).

Complexity:

O(count)

Examples

use stl::*;

let arr = 1..=3;
assert_eq!(arr.lazy_fold_left(0, |x, y| x + y), 6);

fn lazy_fold_right<R, F>(&self, init: R, op: F) -> R
where F: FnMut(Self::Element, R) -> R, Self: BidirectionalCollection, Self::Whole: BidirectionalCollection,

Returns the result of combining elements of given collection using given accumulation operation from right to left.

Postcondition

Result is (e0 + ... + (e(n-1) + (e(n) + init))). where e1, e2, …, en are the references to collection elements, where (a + b) represents op(a, b).

Complexity:

O(count)

Examples

use stl::*;

let arr = 1..=3;
assert_eq!(arr.lazy_fold_right(0, |x, y| x + y), 6);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.