stl

Trait MutableCollection

pub trait MutableCollection: ReorderableCollection
where
    Self::Whole: MutableCollection,
{
    type IterMut<'a>: Iterator<Item = &'a mut Self::Element>
       where Self: 'a;

    // Required methods
    fn at_mut(&mut self, i: &Self::Position) -> &mut Self::Element;
    fn iter_mut_within(
        &mut self,
        from: Self::Position,
        to: Self::Position,
    ) -> Self::IterMut<'_>;

    // Provided method
    fn iter_mut(&mut self) -> Self::IterMut<'_> { ... }
}

Models a collection which supports mutating its element

Required Associated Types

type IterMut<'a>: Iterator<Item = &'a mut Self::Element> where Self: 'a

Type of iterator that can iterate over mutable-ref of elements in self.

Required Methods

fn at_mut(&mut self, i: &Self::Position) -> &mut Self::Element

Mutably Access element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

fn iter_mut_within( &mut self, from: Self::Position, to: Self::Position, ) -> Self::IterMut<'_>

Returns a mutable iterator of self that iterates mutable reference of elements in [from, to) positions.

Precondition

[from, to) represents valid positions in collection.

Provided Methods

fn iter_mut(&mut self) -> Self::IterMut<'_>

Returns a mutable iterator of self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.