stl

Trait ParallelCollectionExt

pub trait ParallelCollectionExt: Collection {

    // Provided methods
    fn first_position_where_par_on<Pred>(
        &self,
        scheduler: &ThreadPool,
        pred: Pred,
    ) -> Option<Self::Position>
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn first_position_where_par<Pred>(
        &self,
        pred: Pred,
    ) -> Option<Self::Position>
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn first_position_of_par_on(
        &self,
        scheduler: &ThreadPool,
        e: &Self::Element,
    ) -> Option<Self::Position>
       where Self::Element: Eq + Sync { ... }
    fn first_position_of_par(&self, e: &Self::Element) -> Option<Self::Position>
       where Self::Element: Eq + Sync { ... }
    fn last_position_where_par_on<Pred>(
        &self,
        scheduler: &ThreadPool,
        pred: Pred,
    ) -> Option<Self::Position>
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn last_position_where_par<Pred>(
        &self,
        pred: Pred,
    ) -> Option<Self::Position>
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn last_position_of_par_on(
        &self,
        scheduler: &ThreadPool,
        e: &Self::Element,
    ) -> Option<Self::Position>
       where Self::Element: Eq + Sync { ... }
    fn last_position_of_par(&self, e: &Self::Element) -> Option<Self::Position>
       where Self::Element: Eq + Sync { ... }
    fn all_satisfy_par_on<Pred>(
        &self,
        scheduler: &ThreadPool,
        pred: Pred,
    ) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn all_satisfy_par<Pred>(&self, pred: Pred) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn any_satisfy_par_on<Pred>(
        &self,
        scheduler: &ThreadPool,
        pred: Pred,
    ) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn any_satisfy_par<Pred>(&self, pred: Pred) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn none_satisfy_par_on<Pred>(
        &self,
        scheduler: &ThreadPool,
        pred: Pred,
    ) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn none_satisfy_par<Pred>(&self, pred: Pred) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }

}

Parallel Algorithms for Collection.

Provided Methods

fn first_position_where_par_on<Pred>( &self, scheduler: &ThreadPool, pred: Pred, ) -> Option<Self::Position>
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Finds position of first element in self satisfying pred. If no such element exists, returns self.end().

Parallel version of first_position_where executing on scheduler.

Complexity

O(n) where n == self.count().

fn first_position_where_par<Pred>(&self, pred: Pred) -> Option<Self::Position>
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Finds position of first element in self satisfying pred. If no such element exists, returns self.end().

Parallel version of first_position_where.

Complexity

O(n) where n == self.count().

fn first_position_of_par_on( &self, scheduler: &ThreadPool, e: &Self::Element, ) -> Option<Self::Position>
where Self::Element: Eq + Sync,

Finds position of first element in self equals e. If no such element exists, returns self.end().

Parallel version of first_position_of executing on scheduler.

Complexity

O(n) where n == self.count().

fn first_position_of_par(&self, e: &Self::Element) -> Option<Self::Position>
where Self::Element: Eq + Sync,

Finds position of first element in self equals e. If no such element exists, returns self.end().

Parallel version of first_position_of.

Complexity

O(n) where n == self.count().

fn last_position_where_par_on<Pred>( &self, scheduler: &ThreadPool, pred: Pred, ) -> Option<Self::Position>
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Finds position of last element in self satisfying pred. If no such element exists, returns self.end().

Parallel version of last_position_where executing on scheduler.

Complexity

O(n) where n == self.count().

fn last_position_where_par<Pred>(&self, pred: Pred) -> Option<Self::Position>
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Finds position of last element in self satisfying pred. If no such element exists, returns self.end().

Parallel version of last_position_where.

Complexity

O(n) where n == self.count().

fn last_position_of_par_on( &self, scheduler: &ThreadPool, e: &Self::Element, ) -> Option<Self::Position>
where Self::Element: Eq + Sync,

Finds position of last element equals e. If no such element exist, return self.end().

Parallel version of last_position_of executing on scheduler.

Complexity

O(n) where n == self.count().

fn last_position_of_par(&self, e: &Self::Element) -> Option<Self::Position>
where Self::Element: Eq + Sync,

Finds position of last element equals e. If no such element exist, return self.end().

Parallel version of last_position_of.

Complexity

O(n) where n == self.count().

fn all_satisfy_par_on<Pred>(&self, scheduler: &ThreadPool, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff all elements in self satisfies pred.

Parallel version of all_satisfy executing on scheduler.

Complexity

O(n) where n == self.count().

fn all_satisfy_par<Pred>(&self, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff all elements in self satisfies pred.

Parallel version of all_satisfy.

Complexity

O(n) where n == self.count().

fn any_satisfy_par_on<Pred>(&self, scheduler: &ThreadPool, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff atleast one element in self satisfies pred.

Parallel version of any_satisfy executing on scheduler.

Complexity

O(n) where n == self.count().

fn any_satisfy_par<Pred>(&self, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff atleast one element in self satisfies pred.

Parallel version of any_satisfy.

Complexity

O(n) where n == self.count().

fn none_satisfy_par_on<Pred>(&self, scheduler: &ThreadPool, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff no element in self satisfies pred.

Parallel version of none_satisfy executing on scheduler.

Complexity

O(n) where n == self.count().

fn none_satisfy_par<Pred>(&self, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff no element in self satisfies pred.

Parallel version of none_satisfy.

Complexity

O(n) where n == self.count().

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.