stl
Trait RandomAccessRange

pub trait RandomAccessRange: BidirectionalRange<Position: Regular + Ord> { }

Models a random access range (similar to array) where jumping to any position from any other position is O(1) operation.

RandomAccessRange is extension of BidirectionalRange.

RandomAccessRange enforces the Position of range should be ordered.

RandomAccessRange doesn’t add any new method but introduces complexity requirements mentioned below. The complexity requirements ensure that any position jump is O(1).

Complexity Requirements

rng.distance(from, to) -> O(1).

rng.after_n(i) -> O(1).

rng.before_n(i) -> O(1).

NOTE: If complexity requirements are not formed any algorithm on RandomAccessRange have undefined behavior.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.