stl

Trait ReorderableCollection

pub trait ReorderableCollection: Collection
where
    Self::SubSequence: ReorderableCollection,
{
    // Required methods
    fn swap_at(&mut self, i: &Self::Position, j: &Self::Position);
    fn slice_mut(
        &mut self,
        from: Self::Position,
        to: Self::Position,
    ) -> SliceMut<'_, Self::SubSequence>;
}

A collection which supports internally reordering its element.

Required Methods

fn swap_at(&mut self, i: &Self::Position, j: &Self::Position)

Swaps element at position i with element at position j.

fn slice_mut( &mut self, from: Self::Position, to: Self::Position, ) -> SliceMut<'_, Self::SubSequence>

A contiguous mutable subrange of self having elements in position [from, to).

The slice share positions with self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.