stl

Trait ReorderableCollectionExt

pub trait ReorderableCollectionExt: ReorderableCollection
where
    Self::Whole: ReorderableCollection,
{
    // Provided methods
    fn all_mut(&mut self) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_mut(&mut self, to: Self::Position) -> SliceMut<'_, Self::Whole> { ... }
    fn suffix_mut(&mut self, from: Self::Position) -> SliceMut<'_, Self::Whole> { ... }
}

Algorithms for ReorderableCollection.

Provided Methods

fn all_mut(&mut self) -> SliceMut<'_, Self::Whole>

Returns mutable slice of the collection covering full collection.

Precondition

Postcondition

Returns slice of the collection covering full collection.

Complexity: O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut s = arr.all_mut();
let start = s.start();
*s.at_mut(&start) = 0;
assert!(s.equals(&[0, 2, 3, 4, 5]));
assert!(arr.equals(&[0, 2, 3, 4, 5]));

fn prefix_mut(&mut self, to: Self::Position) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Postcondition

Returns prefix slice of the collection ending at to exclusive.

Complexity: O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut p = arr.prefix_mut(3);
let start = p.start();
*p.at_mut(&start) = 0;
assert!(p.equals(&[0, 2, 3]));
assert!(arr.equals(&[0, 2, 3, 4, 5]));

fn suffix_mut(&mut self, from: Self::Position) -> SliceMut<'_, Self::Whole>

Returns mutable suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Postcondition

Returns suffix slice of the collection starting from from inclusive.

Complexity: O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut s = arr.suffix_mut(3);
let start = s.start();
*s.at_mut(&start) = 0;
assert!(s.equals(&[0, 5]));
assert!(arr.equals(&[1, 2, 3, 0, 5]));

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.