stl
Trait SemiOutputRange

pub trait SemiOutputRange: ForwardRange {
    // Required method
    fn swap_at(&mut self, i: &Self::Position, j: &Self::Position);
}

Models mutable range that can only be reordered.

SemiOutputRange is extension of ForwardRange.

SemiOutputRange models primitive mutable range that just supports reordering of its elements.

For supporting reordering of elements, swap_at function is required. It allows to swap 2 elements at given positions.

Required Methods

fn swap_at(&mut self, i: &Self::Position, j: &Self::Position)

Swap elements at position i and j

Precondition

i and j is a valid position in self.

Complexity Requirement

O(1)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.