stl

Trait UnsafeSubSequence

pub trait UnsafeSubSequence: Collection {
    // Required methods
    unsafe fn unsafe_at<'a>(&self, i: &Self::Position) -> Self::ElementRef<'a>;
    unsafe fn unsafe_slice(
        &self,
        from: Self::Position,
        to: Self::Position,
    ) -> Self::SubSequence;
    unsafe fn set_start(&mut self, p: Self::Position);
    unsafe fn set_end(&mut self, p: Self::Position);
}

A low-level, unchecked view into contiguous subsequence of a collection.

NOTE: The trait doesn’t enforce usual lifetime tracking. It is intended to be wrapped in a safe view (i.e., Slice).

Required Methods

unsafe fn unsafe_at<'a>(&self, i: &Self::Position) -> Self::ElementRef<'a>

Yields reference to ith element with lifetime 'a.

Safety

It is responsibility of caller to provide such 'a that doesn’t violate memory safety.

unsafe fn unsafe_slice( &self, from: Self::Position, to: Self::Position, ) -> Self::SubSequence

Returns a slice of elements in position range [from, to).

Safety

It is responsibility of caller to use this mechanism to create disjoint slices of appropriate lifetimes.

unsafe fn set_start(&mut self, p: Self::Position)

Set the start position of self to p.

Precondition: p belongs to [self.start(), self.end()].

Safety

It is responsibility of caller to use this mechanism to create disjoint slices of appropriate lifetimes.

unsafe fn set_end(&mut self, p: Self::Position)

Set the end position of self to p.

Precondition: p belongs to [self.start(), self.end()].

Safety

It is responsibility of caller to use this mechanism to create disjoint slices of appropriate lifetimes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.