stl
Module view

Views module

The view module provides collection of views over given range.

A view is just a range that implements View marker trait.

For obtaining view from a range use .view() or .view_mut() function.

use stl::*;

let arr = [1, 2, 3];
let v = arr.view(); // An immutable view of arr.

let mut arr = [1, 2, 3];
let mut v = arr.view_mut(); // A mutable view of arr.

For all views it is mentioned in doc if infix version is supported or not. Or open infix module to see all supported infix views.

For using infix version of supported views use:

use stl::view::infix::*;

Modules

infix

Functions

as_reversed

Returns reverse view of given view.

cycle

Returns a view that is cyclic view of given view.

drop

Returns a view with dropping first n elements of given view.

drop_while

Returns suffix view of given range such that, view starts from the first position p such that p doesn’t satisfy pred.

empty

Returns an empty view.

filter

Creates a filter view which contains only the element that satisfies given predicate.

generate

Returns a view which contains element by repeatedly calling generator, until it returs a None Option.

ints

Returns an infinite sequence of i32s starting from init.

join

Returns a view that flattens a layer of nesting of given view.

map

Creates a view from given view where each element is op(&e).

maybe

Returns a view that contains a single instance of val or is empty.

prefix

Returns prefix of given view, i.e., [view.start(), end).

repeat

Returns a view that contains infinite stream of val.

single

Returns a view that contains an instance of val.

subrange

Returns a subrange of given view from [start, end).

suffix

Returns suffix of given view, i.e., [view.start(), end).

take

Returns a view of first n elements of given view.

take_while

Returns the prefix view of given view such that prefix only contains elements that satisfies the given predicate. It starts excluding elements from given range from first position, element did not satisfy the predicate.

zip_map

Zips the given views using given operation.