stl::view
Function generate

pub fn generate<Generator, T>(
    generator: Generator,
) -> GenerateView<Generator, T>
where
    Generator: FnMut() -> Option<T>,

Returns a view which contains element by repeatedly calling generator, until it returs a None Option.

Precondition

Postcondition

Returns a view which contains element by repeatedly calling generator, until it returs a None Option.

It satisfies following range traits:

InputRange: YES

BoundedRange: NO

ForwardRange: NO

BidirectionalRange: NO

RandomAccessRange: NO

SemiOutputRange: NO

OutputRange: NO

Infix Supported

NO

Examples

use stl::*;
use rng::infix::*;

let mut counter = 0;
let generator = || {
    if counter <= 3 {
        let result = Some(counter);
        counter += 1;
        result
    } else {
        None
    }
};
let arr = view::generate(generator);
assert!(arr.equals(&[0, 1, 2, 3]));