stl::view
Function join

pub fn join<Element, InRange, OutRange>(
    view: OutRange,
) -> JoinView<Element, InRange, OutRange>
where
    InRange: InputRange<Element = Element>,
    for<'a> OutRange: 'a + InputRange<Element = InRange, ElementRef<'a> = &'a InRange> + View,

Returns a view that flattens a layer of nesting of given view.

Precondition

Postcondition

Returns a view that flattens a layer of nesting of given view.

Range traits satisfied:

InputRange: YES

BoundedRange: If OutRange is BoundedRange

ForwardRange: If OutRange, InRange is ForwardRange

BidirectionalRange: If OutRange is BidirectionalRange and InRange is BidirectionalRange + BoundedRange

RandomAccessRange: NO

SemiOutputRange: NO

OutputRange: NO

Infix Supported

YES

TODO

Join view requires, the inner range element ref to strictly be reference to inner range. However, expressing that hits borrow checker current limitation and borrow checker assumes it to be ’static right now. Due to this, test cases are also using static array and makes join unusable at all. Instead find a way to express the same without hitting borrow checker limitation.

Example

use stl::*;
use rng::infix::*;
use view::infix::*;

static ARR: [[i32; 3]; 3] = [[2, 1, 2], [2, 9, 4], [3, 1, 2]];

let v = view::join(ARR.view());
assert_eq!(v.count(&2), 4);

let v = ARR.view().join();
assert_eq!(v.count(&2), 4);