stl::view
Function prefix

pub fn prefix<RangeView>(
    view: RangeView,
    end: RangeView::Position,
) -> PrefixView<RangeView>
where
    RangeView: ForwardRange + View,

Returns prefix of given view, i.e., [view.start(), end).

Precondition

end is a valid position in view.

Postcondition

Returns a view that is prefix [view.start(), end) of given view.

Returned view is:

InputRange: YES

ForwardRange: YES

BoundedRange: YES

BidirectionalRange: If given view is BidirectionalRange

RandomAccessRange: If given view is RandomAccessRange

SemiOutputRange: If given view is SemiOutputRange

OutputRange: If given view is OutputRange

Position type is same as given range Position type

Element type is same as given range Element type

ElementRef type is same as given range ElementRef type

ElementMutRef type is same as given range ElementMutRef type

Infix Supported

YES

Example

use stl::*;
use view::infix::*;
use rng::infix::*;

let mut arr = [(3, 1), (1, 2), (4, 4), (1, 1), (5, 5)];
view::prefix(arr.view_mut(), 4).stable_sort_by(|x, y| x.0 < y.0);
assert_eq!(arr, [(1, 2), (1, 1), (3, 1), (4, 4), (5, 5)]);

let mut arr = [(3, 1), (1, 2), (4, 4), (1, 1), (5, 5)];
arr.view_mut().prefix(4).stable_sort_by(|x, y| x.0 < y.0);
assert_eq!(arr, [(1, 2), (1, 1), (3, 1), (4, 4), (5, 5)]);