Crate stl

STL module

The stl module provides core traits that define regularity, ranges, etc. This module also provides implementation of those traits for many commmon data structures. Currently supported data structures are:

[T;n] (array)

[T] (slice)

Vec<T> (Vec)

Structs

CollectionIterator

Slice

SliceMut

Traits

BidirectionalCollection

Models a bidirectional collection, which can be traversed forward as well as backward.

Collection

Models a multi-pass linear sequence of elements.

CollectionExt

MutableCollection

Models a collection which supports mutating its element

MutableCollectionExt

RandomAccessCollection

Models a random access collection (similar to array) where jumping to any position from any other position is O(1) operation.

RandomAccessCollectionExt

Regular

Any SemiRegular type that is cloneable.

ReorderableCollection

Models a collection which supports internally reordering its element.

ReorderableCollectionExt

SemiRegular

Any type that is movable, destructable and equality comparable.