Crate stl

STL module

The stl module provides formal definition of “Collections” in terms of traits. Based on primitives exposed by traits, the module provides many generic algorithms for types satisfying “Collection” traits.

The module also implements the “Collection” related traits for stdlib data structures. Currently “Collection” traits have been implemented for:

[T;n] (array)

[T] (slice)

Vec<T> (Vec)

Option<T> (Option)

Range<T> (a..b) where T is a signed/unsigned integer type.

RangeInclusive<T> (a..=b) where T is a signed/unsigned integer type.

Structs

CollectionIterator

Generic iterator to iterate over ElementRef of any Collection.

CollectionOfOne

A collection of one element.

EmptyCollection

An empty collection.

LazyCollectionIterator

Generic iterator to iterate over Element of any LazyCollection.

Slice

A contiguous sub-collection of a collection.

SliceMut

A contiguous mutable sub-collection of a mutable collection.

Traits

BidirectionalCollection

Models a bidirectional collection, which can be traversed forward as well as backward.

BinaryPredicate

A BinaryPredicate<T, U> is a “cheap” to copy function that accepts T and U by reference and returns bool.

Collection

Models a multi-pass linear sequence of elements.

CollectionExt

Algorithms for Collection.

LazyCollection

Models a collection whose elements are computed on element access.

LazyCollectionExt

Algorithms for LazyCollection.

MutableCollection

Models a collection which supports mutating its element

MutableCollectionExt

Algorithms for MutableCollection.

Predicate

A Predicate<T> is a “cheap” to copy function that accepts T by reference and returns bool.

RandomAccessCollection

Models a random access collection (similar to array) where jumping to any position from any other position is O(1) operation.

RandomAccessCollectionExt

Algorithms for RandomAccessCollection.

Regular

Any SemiRegular type that is cloneable.

ReorderableCollection

Models a collection which supports internally reordering its element.

ReorderableCollectionExt

Algorithms for ReorderableCollection.

SemiRegular

Any type that is movable, destructable and equality comparable.