Crate stl

STL module

The stl module provides core traits that define regularity, ranges, etc. This module also provides implementation of those traits for many commmon data structures. Currently supported data structures are:

[T;n] (array)

[T] (slice)

Vec<T> (Vec)

In upcoming future more data structures will be supported. If supporting specific data structure would not be possible due to some design choices, alternate data structures would be provided for the same.

Modules

algo

Algorithms module

rng

Ranges module

Traits

BidirectionalRange

Models a bidirectional range, which can be traversed forward as well as backward.

ForwardRange

Models a multi-pass range.

InputRange

Models a single-pass range.

OutputRange

Models a mutable range.

RandomAccessRange

Models a random access range (similar to array) where any position can be reached in O(1).

Regular

As per Stepanov (not exact) Types is semiregular and can be copied. Copy should have equal value as of original object.

SemiRegular

As per Stepanov (not exact) Types can be: