Crate stl

STL module

The stl module provides core traits that define regularity, ranges, etc. This module also provides implementation of those traits for many commmon data structures. Currently supported data structures are:

[T;n] (array)

[T] (slice)

Vec<T> (Vec)

str

String

Re-exports

pub use view::STLMutableViewExt;

pub use view::STLViewExt;

Modules

algo

Algorithms module

rng

Ranges module

view

Views module

Traits

BidirectionalRange

Models a bidirectional range, which can be traversed forward as well as backward.

BoundedRange

Models a bounded range, i.e., the range whose end position is known.

ForwardRange

Models a multi-pass range.

InputRange

Models a single-pass range.

OutputRange

Models a fully mutable range.

RandomAccessRange

Models a random access range (similar to array) where jumping to any position from any other position is O(1) operation.

Regular

SemiRegular + Clonable

SemiOutputRange

Models mutable range that can only be reordered.

SemiRegular

Equality Comparable + Movable + Destructable.

View

Marker trait for a view.