stl::rng
Function unique_copy_by

pub fn unique_copy_by<SrcRange, DestRange, BinaryPred>(
    src: &SrcRange,
    dest: &mut DestRange,
    bi_pred: BinaryPred,
) -> DestRange::Position
where
    SrcRange: InputRange + ?Sized,
    DestRange: OutputRange<Element = SrcRange::Element> + ?Sized,
    SrcRange::Element: Clone,
    BinaryPred: Fn(&SrcRange::Element, &SrcRange::Element) -> bool,

Copies all elements from src to dest with copying only first of adjacent equivalent elements by given equivalence relationship.

Precondition

dest can accomodate copied elements.

binary_pred should follow equivalence relationship, otherwise behavior is undefined.

Postcondition

Copies elements from src with omitting adjacent equivalent elements by bi_pred to dest.

Returns the position of past to last copied element in dest.

Complexity: O(n). Exactly max(0, n - 1) applications of bi_pred.

Where n is number of elements in src.

Infix Supported

NO

Example

use stl::*;
use rng::infix::*;

let src = [(1, 2), (1, 3), (2, 3)];
let mut dest = [(0, 0), (0, 0)];
let i = rng::unique_copy_by(&src, &mut dest, |x, y| x.0 == y.0);
assert_eq!(i, 2);
assert!(dest.equals(&[(1, 2), (2, 3)]));