stl::rng
Module infix

Defines extension traits to support infix version of range algorithms.

For preconditions and postcondition of any algorithm see rng module functions.

Traits

STLAdjacentFindExt

adjacent_find.

STLCountExt

count_if, count.

STLEqualsExt

equals, equals_by.

STLFillExt

fill_value, fill_by.

STLFindExt

find_if, find_if_not, find.

STLOfExt

all_of, any_of, none_of.

STLRemoveExt

remove, remove_if.

STLReplaceExt

replace, replace_if.

STLReverseExt

reverse.

STLRotateExt

rotate.

STLUniqueExt

unique, unique_by.