stl::rng
Module infix

Defines extension traits to support infix version of range algorithms.

For preconditions and postcondition of any algorithm see rng module functions.

Traits

STLAdjacentFindExt

adjacent_find.

STLBinarySearchExt

lower_bound, lower_bound_by, upper_bound, upper_bound_by, equal_range, equal_range_by, binary_search, binary_search_by.

STLCopyExt

copy, copy_if.

STLCountExt

count_if, count.

STLEqualsExt

equals, equals_by.

STLFillExt

fill_value, fill_by.

STLFindExt

find_if, find_if_not, find.

STLForEachInputExt

for_each.

STLForEachOutputExt

for_each_mut.

STLForwardPartitonExt

partition_point.

STLHeapExt

is_heap_until, is_heap_until_by.

STLInputPartitionExt

is_partitioned.

STLMinMaxExt

min_element, min_element_by, max_element, max_element_by, minmax_element, minmax_element_by.

STLNumericBidirExt

fold_right.

STLNumericInputExt

fold_left.

STLOfExt

all_of, any_of, none_of.

STLOutputHeapExt

push_heap, push_heap_by, pop_heap, pop_heap_by, make_heap, make_heap_by, sort_heap, sort_heap_by.

STLOutputPartitionExt

stable_partition.

STLRemoveExt

remove, remove_if.

STLReplaceExt

replace, replace_if.

STLReverseExt

reverse.

STLRotateExt

rotate.

STLSemiOutputBoundedPartitionExt

stable_partition_no_alloc.

STLSemiOutputPartitionExt

partition.

STLSortExt

sort_range, sort_range_by, stable_sort_no_alloc, stable_sort_by_no_alloc, partial_sort, partial_sort_by, nth_element, nth_element_by.

STLSortForwardExt

is_sorted_until, is_sorted_until_by, is_sorted, is_sorted_by, partial_sort, partial_sort_by.

STLSortOutputExt

stable_sort, stable_sort_by.

STLUniqueExt

unique, unique_by.