stl

Struct Slice

pub struct Slice<'a, Whole>
where
    Whole: Collection<Whole = Whole>,
{
    pub from: Whole::Position,
    pub to: Whole::Position,
    /* private fields */
}

A contiguous sub-collection of a collection.

Fields
from: Whole::Position

Start position of slice.

to: Whole::Position

End position of slice.

Implementations

impl<'a, Whole> Slice<'a, Whole>
where Whole: Collection<Whole = Whole>,

pub fn new( collection: &'a Whole, from: Whole::Position, to: Whole::Position, ) -> Self

Creates a new instance of slice with given collection and position range.

pub fn whole(&self) -> &'a Whole

Returns the reference to whole collection.

pub fn at(&self, i: &Whole::Position) -> Whole::ElementRef<'a>

Access element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

impl<Whole> Slice<'_, Whole>
where Whole: Collection<Whole = Whole>,

Dropping algorithms

pub fn drop_first(&mut self) -> bool

Removes the first element if non-empty and returns true; returns false otherwise.

Example

use stl::*;

let arr = [1, 2, 3];
let mut s = arr.full();
assert!(s.drop_first());
assert!(s.equals(&[2, 3]));

pub fn drop_last(&mut self) -> bool
where Whole: BidirectionalCollection,

Removes the last element if non-empty and returns true; returns false otherwise.

Example

use stl::*;

let arr = [1, 2, 3];
let mut s = arr.full();
assert!(s.drop_last());
assert!(s.equals(&[1, 2]));

pub fn drop_prefix(&mut self, max_length: usize)

Drops prefix upto specified maximum length.

Postcondition

If max_length > self.count(), make self empty.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise, where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
s.drop_prefix(2);
assert!(s.equals(&[2, 3]));

pub fn drop_prefix_upto(&mut self, position: Whole::Position)

Drops the prefix of slice upto given position.

Precondition

position is valid position in self.

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
s.drop_prefix_upto(2);
assert!(s.equals(&[2, 3]));

pub fn drop_prefix_through(&mut self, position: Whole::Position)

Drops the prefix of slice till and including given position.

Precondition

position is valid position in self.

position != self.end()

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
s.drop_prefix_through(2);
assert!(s.equals(&[3]));

pub fn drop_while<Pred>(&mut self, predicate: Pred)
where Pred: FnMut(&Whole::Element) -> bool,

Drops the element of self while the elements satisfy given predicate.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
s.drop_while(|x| *x < 2);
assert!(s.equals(&[2, 3]));

pub fn drop_suffix(&mut self, max_length: usize)

Drops suffix upto specified maximum length.

Postcondition

If max_length > self.count(), make self empty.

Complexity

O(n), where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
s.drop_suffix(2);
assert!(s.equals(&[0, 1]));

pub fn drop_suffix_from(&mut self, position: Whole::Position)

Drops suffix from given position.

Precondition

position is a valid position in self.

Complexity

O(n), where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
s.drop_suffix_from(2);
assert!(s.equals(&[0, 1]));

impl<'a, Whole> Slice<'a, Whole>
where Whole: Collection<Whole = Whole>,

Pop algorithms.

pub fn pop_first(&mut self) -> Option<Whole::ElementRef<'a>>

Removes and returns the first element if non-empty; returns None otherwise.

Example

use stl::*;

let arr = [1, 2, 3];
let mut s = arr.full();
let first = s.pop_first().unwrap();
assert_eq!(*first, 1);
assert!(s.equals(&[2, 3]));

pub fn pop_first_with_pos( &mut self, ) -> Option<(Whole::Position, Whole::ElementRef<'a>)>

Removes and returns the first element and its position if non-empty; returns None otherwise.

Example

use stl::*;

let arr = [1, 2, 3];
let mut s = arr.full();
let (first_pos, first) = s.pop_first_with_pos().unwrap();
assert_eq!(first_pos, 0);
assert_eq!(*first, 1);
assert!(s.equals(&[2, 3]));

pub fn pop_last(&mut self) -> Option<Whole::ElementRef<'a>>
where Whole: BidirectionalCollection,

Removes and returns the last element if non-empty; returns None otherwise.

Example

use stl::*;

let arr = [1, 2, 3];
let mut s = arr.full();
let last = s.pop_last().unwrap();
assert_eq!(*last, 3);
assert!(s.equals(&[1, 2]));

pub fn pop_last_with_pos( &mut self, ) -> Option<(Whole::Position, Whole::ElementRef<'a>)>
where Whole: BidirectionalCollection,

Removes and returns the last element and its position if non-empty; returns None otherwise.

Example

use stl::*;

let arr = [1, 2, 3];
let mut s = arr.full();
let (last_pos, last) = s.pop_last_with_pos().unwrap();
assert_eq!(last_pos, 2);
assert_eq!(*last, 3);
assert!(s.equals(&[1, 2]));

pub fn pop_prefix(&mut self, max_length: usize) -> Self

Removes and returns prefix upto specified maximum length.

Postcondition

If max_length > self.count(), make self empty and return the full slice as result.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise, where n == self.count().

Examples

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
let prefix = s.pop_prefix(2);
assert!(prefix.equals(&[0, 1]));
assert!(s.equals(&[2, 3]));

pub fn pop_prefix_upto(&mut self, position: Whole::Position) -> Self

Removes and returns the prefix slice upto given position.

Precondition

position is valid position in self.

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
let prefix = s.pop_prefix_upto(2);
assert!(prefix.equals(&[0, 1]));
assert!(s.equals(&[2, 3]));

pub fn pop_prefix_through(&mut self, position: Whole::Position) -> Self

Removes and returns the prefix till and including given position.

Precondition

position is valid position in self.

position != self.end()

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
let prefix = s.pop_prefix_through(2);
assert!(prefix.equals(&[0, 1, 2]));
assert!(s.equals(&[3]));

pub fn pop_while<Pred>(&mut self, predicate: Pred) -> Self
where Pred: FnMut(&Whole::Element) -> bool,

Removes and returns the element of self till the first element satisfies predicate as a slice.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
let prefix = s.pop_while(|x| *x < 2);
assert!(prefix.equals(&[0, 1]));
assert!(s.equals(&[2, 3]));

pub fn pop_suffix(&mut self, max_length: usize) -> Self

Removes and returns suffix upto specified maximum length.

Postcondition

If max_length > self.count(), make self empty and returns the full slice as suffix.

Complexity

O(n), where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
let suffix = s.pop_suffix(2);
assert!(s.equals(&[0, 1]));
assert!(suffix.equals(&[2, 3]));

pub fn pop_suffix_from(&mut self, position: Whole::Position) -> Self

Removes and returns suffix from given position.

Precondition

position is a valid position in self.

Complexity

O(n), where n == self.count().

Example

use stl::*;

let arr = [0, 1, 2, 3];
let mut s = arr.full();
let suffix = s.pop_suffix_from(2);
assert!(s.equals(&[0, 1]));
assert!(suffix.equals(&[2, 3]));

impl<'a, Whole> Slice<'a, Whole>
where Whole: LazyCollection<Whole = Whole>,

Pop algorithms for LazyCollection.

pub fn lazy_pop_first(&mut self) -> Option<Whole::Element>

Removes and returns the lazily computed first element if non-empty; returns None otherwise.

Example

use stl::*;

let arr = 1..=3;
let mut s = arr.full();
let first = s.lazy_pop_first().unwrap();
assert_eq!(first, 1);
assert!(s.equals(&[2, 3]));

pub fn lazy_pop_first_with_pos( &mut self, ) -> Option<(Whole::Position, Whole::Element)>

Removes and returns the lazily computed first element and its position if non-empty; returns None otherwise.

Example

use stl::*;

let arr = 1..=3;
let mut s = arr.full();
let (first_pos, first) = s.lazy_pop_first_with_pos().unwrap();
assert_eq!(first_pos, 1);
assert_eq!(first, 1);
assert!(s.equals(&[2, 3]));

pub fn lazy_pop_last(&mut self) -> Option<Whole::Element>
where Whole: BidirectionalCollection,

Removes and returns the lazily computed last element if non-empty; returns None otherwise.

Example

use stl::*;

let arr = 1..=3;
let mut s = arr.full();
let last = s.lazy_pop_last().unwrap();
assert_eq!(last, 3);
assert!(s.equals(&[1, 2]));

pub fn lazy_pop_last_with_pos( &mut self, ) -> Option<(Whole::Position, Whole::Element)>
where Whole: BidirectionalCollection,

Removes and returns the lazily computed last element and its position if non-empty; returns None otherwise.

Example

use stl::*;

let arr = 1..=3;
let mut s = arr.full();
let (last_pos, last) = s.lazy_pop_last_with_pos().unwrap();
assert_eq!(last_pos, 3);
assert_eq!(last, 3);
assert!(s.equals(&[1, 2]));

impl<'a, Whole> Slice<'a, Whole>
where Whole: Collection<Whole = Whole>,

Splitting algorithms.

pub fn split_at(self, position: Whole::Position) -> (Self, Self)

Returns two disjoint slices of self split at the given position.

Examples

use stl::*;

let arr = [0, 1, 2, 3, 4];
let (s1, s2) = arr.full().split_at(2);
assert!(s1.equals(&[0, 1]));
assert!(s2.equals(&[2, 3, 4]));

pub fn split_after(self, position: Whole::Position) -> (Self, Self)

Returns two disjoint slices of self, split immediately after the given position.

Precondition

position != self.end().

Examples

use stl::*;

let arr = [0, 1, 2, 3, 4];
let (s1, s2) = arr.full().split_after(2);
assert!(s1.equals(&[0, 1, 2]));
assert!(s2.equals(&[3, 4]));

pub fn split_where<Pred>( self, pred: Pred, ) -> SplitWhereIterator<'a, Whole, Pred>
where Pred: FnMut(&Whole::Element) -> bool, Self: Sized,

Splits self into slices separated by elements that satisfy pred by returning an Iterator of slices.

Example

use stl::*;

let arr = [1, 3, 5, 2, 2, 3, 4, 5, 5];

// Store sum of each split.
let mut res = vec![];
arr.full()
   .split_where(|x| x % 2 == 0)
   .for_each(|s| res.push(s.iter().sum::<i32>()));
assert_eq!(res, vec![9, 0, 3, 10]);

pub fn split_evenly_in_with_min_size( self, max_slices: usize, min_size: usize, ) -> SplitEvenlyIterator<'a, Whole>

Splits self into at max max_slices slices with each slice being of at min size of min_size by returning an Iterator of slices.

Precondition

max_slices > 0,

Postcondition

If splitting exactly evenly is not possible, then slices on start would have bigger size than slices at end, still maintaining as even splitting as possible.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr.full()
    .split_evenly_in_with_min_size(3, 2)
    .map(|s| s.iter().copied().collect())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);

pub fn split_evenly_in( self, num_slices: usize, ) -> SplitEvenlyIterator<'a, Whole>

Splits self into num_slices slices by returning an Iterator of slices.

Precondition

num_slices > 0,

Postcondition

If splitting exactly evenly is not possible, then slices on start would have bigger size than slices at end, still maintaining as even splitting as possible.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr.full()
    .split_evenly_in(3)
    .map(|s| s.iter().copied().collect())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);