stl

Struct Slice

pub struct Slice<'a, Whole>
where
    Whole: Collection<Whole = Whole>,
{ /* private fields */ }

A contiguous sub-collection of a collection.

Implementations

impl<'a, Whole> Slice<'a, Whole>
where Whole: Collection<Whole = Whole>,

pub fn new( collection: &'a Whole, from: Whole::Position, to: Whole::Position, ) -> Self

Creates a new instance of slice with given collection and position range.

impl<Whole> Slice<'_, Whole>
where Whole: Collection<Whole = Whole>,

Dropping algorithms

pub fn drop_first(&mut self) -> bool

Removes the first element if non-empty and returns true; returns false otherwise.

Complexity

O(1).

pub fn drop_last(&mut self) -> bool
where Whole: BidirectionalCollection,

Removes the last element if non-empty and returns true; returns false otherwise.

Complexity

O(1).

pub fn drop(&mut self, n: usize)

Removes first n elements from self if self has atleast n elements; otherwise make self empty.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise.

pub fn drop_prefix_upto(&mut self, position: Whole::Position)

Removes the subsequence from start position of self up to, but not including p.

Complexity

O(1).

pub fn drop_prefix_through(&mut self, position: Whole::Position)

Removes the subsequence from start position of self through (including) p.

Complexity

O(1).

pub fn drop_while<Predicate>(&mut self, p: Predicate)
where Predicate: FnMut(&Whole::Element) -> bool,

Removes the longest prefix of self whose elements satisfy p.

Complexity

Atmost self.count() applications of p.

pub fn drop_suffix(&mut self, n: usize)

Removes last n elements from self. If self has less than n elements, make it empty.

Complexity

O(1) for RandomAccessCollection;

O(self.count()) otherwise.

pub fn drop_suffix_from(&mut self, p: Whole::Position)

Removes last elements from self starting from p.

Complexity

O(1).

impl<'a, Whole> Slice<'a, Whole>
where Whole: Collection<Whole = Whole>,

Pop algorithms.

pub fn pop_first(&mut self) -> Option<Whole::ElementRef<'a>>

Removes and yields the first element if non-empty; returns None otherwise.

Complexity:

O(1).

pub fn pop_last(&mut self) -> Option<Whole::ElementRef<'a>>
where Whole: BidirectionalCollection,

Removes and yields the last element if non-empty; returns None otherwise.

Complexity

O(1).

pub fn pop(&mut self, n: usize) -> Self

Removes and returns subsequence of first n elements in self; If self has less than n elements, make self empty and return all elements of self.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise.

pub fn pop_prefix_upto(&mut self, p: Whole::Position) -> Self

Removes and returns the subsequence from start position of self up to, but not including p.

Complexity

O(1).

pub fn pop_prefix_through(&mut self, p: Whole::Position) -> Self

Removes and returns the subsequence from start position of self through (including) p.

Complexity

O(1).

pub fn pop_while<Predicate>(&mut self, p: Predicate) -> Self
where Predicate: FnMut(&Whole::Element) -> bool,

Removes and returns the longest prefix of self whose elements satisfy p.

Complexity

Atmost self.count() applications of p.

pub fn pop_end(&mut self, n: usize) -> Self

Removes and returns subsequence of last n elements from self. If self has less than n elements, make it empty and returns subsequence of all elements.

Complexity

O(n).

pub fn pop_suffix_from(&mut self, p: Whole::Position) -> Self

Removes and returns subsequence of last elements from self starting from p.

Complexity

O(1).

impl<'a, Whole> Slice<'a, Whole>
where Whole: Collection<Whole = Whole>,

Splitting algorithms.

pub fn split_at(self, p: Whole::Position) -> (Self, Self)

Splits self into two subsequences at position p:

the left part contains elements before p,

the right part contains elements starting at p.

Complexity

O(1).

pub fn split_after(self, position: Whole::Position) -> (Self, Self)

Splits self into two subsequences at position p:

the left part contains elements before p,

the right part contains elements starting at p.

Complexity

O(1).

pub fn split_where<Pred>( self, pred: Pred, ) -> SplitWhereIterator<'a, Whole, Pred>
where Pred: FnMut(&Whole::Element) -> bool, Self: Sized,

Returns an iterator over subsequences of self, split at elements where p returns true.

Note

Consecutive elements for which p returns true produce empty subsequences.

Complexity

O(self.count()).

Example

use stl::*;

let mut arr = [1, 3, 5, 2, 2, 2, 3, 4, 5, 7];
let v: Vec<_> =
  arr.full_mut()
     .split_where(|x| x % 2 == 0)
     .map(|s| s.to_vec())
     .collect();
assert_eq!(v, vec![vec![1, 3, 5], vec![], vec![], vec![3], vec![5, 7]]);

pub fn split_evenly_in_with_min_size( self, n: usize, min_size: usize, ) -> SplitEvenlyIterator<'a, Whole>

Returns an iterator over at most n subsequences of self, each of size at least min_size, splitting as evenly as possible.

If the elements cannot be divided evenly, the earlier subsequences are one element larger than the later ones.

Precondition

n > 0.

Complexity

O(1) for RandomAccessCollection;

O(self.count()) otherwise.

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr.full_mut()
    .split_evenly_in_with_min_size(3, 2)
    .map(|s| s.to_vec())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);

pub fn split_evenly_in(self, n: usize) -> SplitEvenlyIterator<'a, Whole>

Returns an iterator over n subsequences of self, split as evenly as possible.

If the elements cannot be divided evenly, the earlier subsequences are one element larger than the later ones.

Precondition

n > 0.

Complexity

O(1) for RandomAccessCollection;

O(self.count()) otherwise.

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr.full_mut()
    .split_evenly_in(3)
    .map(|s| s.to_vec())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);