stl

Trait BidirectionalCollection

pub trait BidirectionalCollection: Collection
where
    Self::Whole: BidirectionalCollection,
{
    // Required method
    fn form_prior(&self, position: &mut Self::Position);

    // Provided methods
    fn form_prior_n(&self, position: &mut Self::Position, n: usize) { ... }
    fn form_prior_n_limited_by(
        &self,
        position: &mut Self::Position,
        n: usize,
        limit: Self::Position,
    ) -> bool { ... }
    fn prior(&self, position: Self::Position) -> Self::Position { ... }
    fn prior_n(&self, position: Self::Position, n: usize) -> Self::Position { ... }
    fn prior_n_limited_by(
        &self,
        position: Self::Position,
        n: usize,
        limit: Self::Position,
    ) -> Option<Self::Position> { ... }
}

Models a bidirectional collection, which can be traversed forward as well as backward.

Required Methods

fn form_prior(&self, position: &mut Self::Position)

Mutates the given position to position just before position.

Precondition

position != start()

Provided Methods

fn form_prior_n(&self, position: &mut Self::Position, n: usize)

Mutates the given position to nth position before position.

Precondition

There are n valid positions in self before position.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn form_prior_n_limited_by( &self, position: &mut Self::Position, n: usize, limit: Self::Position, ) -> bool

Decrements given position by n, or so that it equals the given limit.

Precondition

limit is a valid position in self.

limit should be decrement-reachable from position.

Postcondition

Returns true if position has been decremented exactly by n without going beyond limit, otherwise returns false.

When the return value is false, the value position is equal to limit.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn prior(&self, position: Self::Position) -> Self::Position

Returns position immediately before position

Precondition

position != start()

fn prior_n(&self, position: Self::Position, n: usize) -> Self::Position

Returns nth position before position

Precondition

self has n valid positions before position.

Complexity

n applications of before.

fn prior_n_limited_by( &self, position: Self::Position, n: usize, limit: Self::Position, ) -> Option<Self::Position>

Returns nth position before given position, unless n is beyond limit.

Precondition

limit is a valid position in self.

limit should be increment-reachable from position.

Postcondition

Returns nth position before given position, unless n is beyond limit.

Otherwise, returns None.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.