stl

Trait Collection

pub trait Collection {
    type Position: Regular;
    type Element;
    type Whole: Collection<Position = Self::Position, Element = Self::Element, Whole = Self::Whole>;

    // Required methods
    fn start(&self) -> Self::Position;
    fn end(&self) -> Self::Position;
    fn form_next(&self, position: &mut Self::Position);
    fn at(&self, i: &Self::Position) -> &Self::Element;
    fn slice(
        &self,
        from: Self::Position,
        to: Self::Position,
    ) -> Slice<'_, Self::Whole>;

    // Provided methods
    fn form_next_n(&self, position: &mut Self::Position, n: usize) { ... }
    fn next(&self, position: Self::Position) -> Self::Position { ... }
    fn next_n(&self, position: Self::Position, n: usize) -> Self::Position { ... }
    fn distance(&self, from: Self::Position, to: Self::Position) -> usize { ... }
    fn count(&self) -> usize { ... }
    fn underestimated_count(&self) -> usize { ... }
    fn iter(&self) -> CollectionIterator<'_, Self::Whole>  { ... }
}

Models a multi-pass linear sequence of elements.

Representation:

  _ _ _ _ _ _

  ^            ^
  |            |
start   -->   end

Required Associated Types

type Position: Regular

Type of positions in the collection.

type Element

Type of element in the collection.

type Whole: Collection<Position = Self::Position, Element = Self::Element, Whole = Self::Whole>

Type representing whole collection. i.e., Self == Slice<W> ? W : Self

Required Methods

fn start(&self) -> Self::Position

Returns the position of first element in self, or if self is empty then start() == end()

fn end(&self) -> Self::Position

Returns the position just after last element in collection.

fn form_next(&self, position: &mut Self::Position)

Mutates given position to next position in collection.

Precondition

position != end()

fn at(&self, i: &Self::Position) -> &Self::Element

Access element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

fn slice( &self, from: Self::Position, to: Self::Position, ) -> Slice<'_, Self::Whole>

Returns slice of collection in positions [from, to).

Precondition

[from, to) represents valid positions in collection.

Provided Methods

fn form_next_n(&self, position: &mut Self::Position, n: usize)

Mutates given position to nth position after current position.

Precondition

There are n valid positions in self after position.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn next(&self, position: Self::Position) -> Self::Position

Returns position immediately after position.

Precondition

position != end()

fn next_n(&self, position: Self::Position, n: usize) -> Self::Position

Returns nth position after position.

Precondition

There are n valid positions in self after position.

Complexity

O(1) for RandomAccessCollection; O(n) otherwise.

fn distance(&self, from: Self::Position, to: Self::Position) -> usize

Returns number of elements in [from, to).

Precondition

[from, to) represents valid positions in the collection.

Complexity

O(1) if RandomAccessCollection; O(n) otherwise.

fn count(&self) -> usize

Returns number of elements in collection.

Complexity

O(1) if RandomAccessCollection; O(n) otherwise.

fn underestimated_count(&self) -> usize

Returns count less than or equal to number of elements in collection.

Complexity

O(1) if RandomAccessCollection; O(n) otherwise.

fn iter(&self) -> CollectionIterator<'_, Self::Whole>

Returns a non-consuming iterator that iterates over &Self::Element.