stl

Trait CollectionExt

pub trait CollectionExt: Collection {

    // Provided methods
    fn iter(&self) -> CollectionIterator<'_, Self::Whole>  { ... }
    fn for_each<F>(&self, f: F)
       where F: FnMut(&Self::Element) { ... }
    fn first(&self) -> Option<<Self as Collection>::ElementRef<'_>> { ... }
    fn all(&self) -> Slice<'_, Self::Whole> { ... }
    fn prefix(&self, to: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn suffix(&self, from: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn equals_by<OtherCollection, F>(
        &self,
        other: &OtherCollection,
        bi_pred: F,
    ) -> bool
       where OtherCollection: Collection,
             F: BinaryPredicate<Self::Element, OtherCollection::Element> { ... }
    fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
       where OtherCollection: Collection<Element = Self::Element>,
             Self::Element: Eq { ... }
    fn first_position_where<Pred>(&self, pred: Pred) -> Self::Position
       where Pred: Predicate<Self::Element> { ... }
    fn first_position_of(&self, e: &Self::Element) -> Self::Position
       where Self::Element: Eq { ... }
    fn last_position_where<Pred>(&self, pred: Pred) -> Self::Position
       where Pred: Predicate<Self::Element> { ... }
    fn last_position_of(&self, e: &Self::Element) -> Self::Position
       where Self::Element: Eq { ... }
    fn all_satisfy<Pred: Predicate<Self::Element>>(&self, pred: Pred) -> bool { ... }
    fn any_satisfy<Pred: Predicate<Self::Element>>(&self, pred: Pred) -> bool { ... }
    fn none_satisfy<Pred: Predicate<Self::Element>>(&self, pred: Pred) -> bool { ... }
    fn count_where<Pred>(&self, pred: Pred) -> usize
       where Pred: Predicate<Self::Element> { ... }
    fn count_of(&self, e: &Self::Element) -> usize
       where Self::Element: Eq { ... }
    fn partition_point<F>(&self, belongs_in_second_half: F) -> Self::Position
       where F: Predicate<Self::Element> { ... }

}

Algorithms for Collection.

Provided Methods

fn iter(&self) -> CollectionIterator<'_, Self::Whole>

Returns a non-consuming iterator that iterates over &Self::Element.

fn for_each<F>(&self, f: F)
where F: FnMut(&Self::Element),

Applies f to each element of collection.

Complexity:

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let mut sum = 0;
arr.for_each(|x| sum = sum + x);
assert_eq!(sum, 6);

fn first(&self) -> Option<<Self as Collection>::ElementRef<'_>>

Returns the first element, or nil if self is empty.

fn all(&self) -> Slice<'_, Self::Whole>

Returns slice of the collection covering full collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.all();
assert!(s.equals(&[1, 2, 3, 4, 5]));

fn prefix(&self, to: Self::Position) -> Slice<'_, Self::Whole>

Returns prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let p = arr.prefix(3);
assert!(p.equals(&[1, 2, 3]));

fn suffix(&self, from: Self::Position) -> Slice<'_, Self::Whole>

Returns suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.suffix(3);
assert!(s.equals(&[4, 5]));

fn equals_by<OtherCollection, F>( &self, other: &OtherCollection, bi_pred: F, ) -> bool
where OtherCollection: Collection, F: BinaryPredicate<Self::Element, OtherCollection::Element>,

Returns true if elements of self is equivalent to elements of other by given relation bi_pred.

Postcondition

Returns true if elements of self is equivalent to other by given relation bi_pred.

If self and other have different number of elements, then return false.

Complexity

O(min(m, n)) where

m == self.count()

n == other.count()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [2, 3, 4];
assert!(arr1.equals_by(&arr2, |x, y| *y == x + 1));

fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
where OtherCollection: Collection<Element = Self::Element>, Self::Element: Eq,

Returns true if elements of self is equal to elements of other.

Postcondition

Returns true if elements of self is equal to elements of other.

If self and other have different number of elements, then return false.

Complexity

O(min(m, n)) where

m == self.count()

n == other.count()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [1, 2, 3];
assert!(arr1.equals(&arr2));

fn first_position_where<Pred>(&self, pred: Pred) -> Self::Position
where Pred: Predicate<Self::Element>,

Finds position of first element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let i = arr.first_position_where(|x| *x == 3);
assert_eq!(i, 2);

fn first_position_of(&self, e: &Self::Element) -> Self::Position
where Self::Element: Eq,

Finds position of first element in self equals e. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 3];
let i = arr.first_position_of(&3);
assert_eq!(i, 2);

fn last_position_where<Pred>(&self, pred: Pred) -> Self::Position
where Pred: Predicate<Self::Element>,

Finds position of last element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4];
let i = arr.last_position_where(|x| x % 2 == 1);
assert_eq!(i, 2);

fn last_position_of(&self, e: &Self::Element) -> Self::Position
where Self::Element: Eq,

Finds position of last element equals e. If no such element exist, return self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 3];
let i = arr.last_position_of(&3);
assert_eq!(i, 2);

fn all_satisfy<Pred: Predicate<Self::Element>>(&self, pred: Pred) -> bool

Returns true if all element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 5];
assert!(arr.all_satisfy(|x| x % 2 == 1));

fn any_satisfy<Pred: Predicate<Self::Element>>(&self, pred: Pred) -> bool

Returns true if atleast one element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 5];
assert!(arr.any_satisfy(|x| x % 2 == 1));

fn none_satisfy<Pred: Predicate<Self::Element>>(&self, pred: Pred) -> bool

Returns true if none of elements in self satisfy pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [2, 4, 6];
assert!(arr.none_satisfy(|x| x % 2 == 1));

fn count_where<Pred>(&self, pred: Pred) -> usize
where Pred: Predicate<Self::Element>,

Returns number of elements in self satisfying pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let n = arr.count_where(|x| x % 2 == 1);
assert_eq!(n, 2);

fn count_of(&self, e: &Self::Element) -> usize
where Self::Element: Eq,

Returns number of elements in self equals e.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [3, 0, 3];
let n = arr.count_of(&3);
assert_eq!(n, 2);

fn partition_point<F>(&self, belongs_in_second_half: F) -> Self::Position
where F: Predicate<Self::Element>,

Returns position of first element of collection for which predicate returns false.

Precondition

The collection should be already partitioned wrt predicate i.e, there exist a position i such that predicate is true for every element of self.prefix(i) and predicate is false for every element of self.suffix(i).

Complexity

O(log n) for RandomAccessCollection, O(n) otherwise; where n == self.count().

O(log n) application of belongs_in_second_half.

Example

use stl::*;

let arr = [1, 3, 5, 2, 4];
let i = arr.partition_point(|x| x % 2 == 0);
assert_eq!(i, 3);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.