stl

Trait CollectionExt

pub trait CollectionExt: Collection {

    // Provided methods
    fn for_each<F>(&self, f: F)
       where F: FnMut(&Self::Element) { ... }
    fn first(&self) -> Option<<Self as Collection>::ElementRef<'_>> { ... }
    fn full(&self) -> Slice<'_, Self::Whole> { ... }
    fn prefix(&self, max_length: usize) -> Slice<'_, Self::Whole> { ... }
    fn prefix_upto(&self, pos: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn prefix_through(&self, pos: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn prefix_while<F: FnMut(&Self::Element) -> bool>(
        &self,
        predicate: F,
    ) -> Slice<'_, Self::Whole> { ... }
    fn drop_while<F>(&self, predicate: F) -> Slice<'_, Self::Whole>
       where F: FnMut(&Self::Element) -> bool { ... }
    fn drop(&self, count: usize) -> Slice<'_, Self::Whole> { ... }
    fn drop_end(&self, count: usize) -> Slice<'_, Self::Whole> { ... }
    fn suffix(&self, max_length: usize) -> Slice<'_, Self::Whole> { ... }
    fn suffix_from(&self, from: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn equals_by<OtherCollection, F>(
        &self,
        other: &OtherCollection,
        bi_pred: F,
    ) -> bool
       where OtherCollection: Collection,
             F: FnMut(&Self::Element, &OtherCollection::Element) -> bool { ... }
    fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
       where OtherCollection: Collection<Element = Self::Element>,
             Self::Element: Eq { ... }
    fn first_position_where<Pred>(&self, pred: Pred) -> Self::Position
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn first_position_of(&self, e: &Self::Element) -> Self::Position
       where Self::Element: Eq { ... }
    fn last_position_where<Pred>(&self, pred: Pred) -> Self::Position
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn last_position_of(&self, e: &Self::Element) -> Self::Position
       where Self::Element: Eq { ... }
    fn all_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn any_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn none_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn count_where<Pred>(&self, pred: Pred) -> usize
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn count_of(&self, e: &Self::Element) -> usize
       where Self::Element: Eq { ... }
    fn partition_point<F>(&self, belongs_in_second_half: F) -> Self::Position
       where F: FnMut(&Self::Element) -> bool { ... }
    fn fold_left<R, F>(&self, init: R, op: F) -> R
       where F: FnMut(R, &Self::Element) -> R { ... }

}

Algorithms for Collection.

Provided Methods

fn for_each<F>(&self, f: F)
where F: FnMut(&Self::Element),

Applies f to each element of collection.

Complexity:

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let mut sum = 0;
arr.for_each(|x| sum = sum + x);
assert_eq!(sum, 6);

fn first(&self) -> Option<<Self as Collection>::ElementRef<'_>>

Returns the first element, or nil if self is empty.

fn full(&self) -> Slice<'_, Self::Whole>

Returns slice of the collection covering full collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.full();
assert!(s.equals(&[1, 2, 3, 4, 5]));

fn prefix(&self, max_length: usize) -> Slice<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the initial elements of collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(k) where k is the number of elements in resultant slice.

Example

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.prefix(3);
assert!(s.equals(&[1, 2, 3]));

fn prefix_upto(&self, pos: Self::Position) -> Slice<'_, Self::Whole>

Returns prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let p = arr.prefix_upto(3);
assert!(p.equals(&[1, 2, 3]));

fn prefix_through(&self, pos: Self::Position) -> Slice<'_, Self::Whole>

Returns prefix slice of the collection ending at to inclusive.

Precondition

to is a valid position in the collection and to != self.end().

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let p = arr.prefix_through(3);
assert!(p.equals(&[1, 2, 3, 4]));

fn prefix_while<F: FnMut(&Self::Element) -> bool>( &self, predicate: F, ) -> Slice<'_, Self::Whole>

Returns prefix slice containing the initial elements until predicate returns false and skipping the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let arr = [1, 3, 5, 2, 7];
let p = arr.prefix_while(|x| x % 2 == 1);
assert!(p.equals(&[1, 3, 5]));

fn drop_while<F>(&self, predicate: F) -> Slice<'_, Self::Whole>
where F: FnMut(&Self::Element) -> bool,

Returns a slice by skipping initial elements while predicate returns true, returning the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let arr = [1, 3, 5, 2, 4, 7];
let s = arr.drop_while(|x| x % 2 == 1);
assert!(s.equals(&[2, 4, 7]));

fn drop(&self, count: usize) -> Slice<'_, Self::Whole>

Returns a slice containing all but the given number of initial elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(k) otherwise, where k is number of elements to drop from beginning of collection.

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.drop(3);
assert!(s.equals(&[4, 5]));

fn drop_end(&self, count: usize) -> Slice<'_, Self::Whole>

Returns a slice containing all but the given number of final elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise, where n == self.count().

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.drop_end(3);
assert!(s.equals(&[1, 2]));

fn suffix(&self, max_length: usize) -> Slice<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the final elements of the collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.suffix(3);
assert!(s.equals(&[3, 4, 5]));

fn suffix_from(&self, from: Self::Position) -> Slice<'_, Self::Whole>

Returns suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.suffix_from(3);
assert!(s.equals(&[4, 5]));

fn equals_by<OtherCollection, F>( &self, other: &OtherCollection, bi_pred: F, ) -> bool
where OtherCollection: Collection, F: FnMut(&Self::Element, &OtherCollection::Element) -> bool,

Returns true if elements of self is equivalent to elements of other by given relation bi_pred.

Postcondition

Returns true if elements of self is equivalent to other by given relation bi_pred.

If self and other have different number of elements, then return false.

Complexity

O(min(m, n)) where

m == self.count()

n == other.count()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [2, 3, 4];
assert!(arr1.equals_by(&arr2, |x, y| *y == x + 1));

fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
where OtherCollection: Collection<Element = Self::Element>, Self::Element: Eq,

Returns true if elements of self is equal to elements of other.

Postcondition

Returns true if elements of self is equal to elements of other.

If self and other have different number of elements, then return false.

Complexity

O(min(m, n)) where

m == self.count()

n == other.count()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [1, 2, 3];
assert!(arr1.equals(&arr2));

fn first_position_where<Pred>(&self, pred: Pred) -> Self::Position
where Pred: FnMut(&Self::Element) -> bool,

Finds position of first element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let i = arr.first_position_where(|x| *x == 3);
assert_eq!(i, 2);

fn first_position_of(&self, e: &Self::Element) -> Self::Position
where Self::Element: Eq,

Finds position of first element in self equals e. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 3];
let i = arr.first_position_of(&3);
assert_eq!(i, 2);

fn last_position_where<Pred>(&self, pred: Pred) -> Self::Position
where Pred: FnMut(&Self::Element) -> bool,

Finds position of last element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4];
let i = arr.last_position_where(|x| x % 2 == 1);
assert_eq!(i, 2);

fn last_position_of(&self, e: &Self::Element) -> Self::Position
where Self::Element: Eq,

Finds position of last element equals e. If no such element exist, return self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 3];
let i = arr.last_position_of(&3);
assert_eq!(i, 2);

fn all_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: FnMut(&Self::Element) -> bool,

Returns true if all element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 5];
assert!(arr.all_satisfy(|x| x % 2 == 1));

fn any_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: FnMut(&Self::Element) -> bool,

Returns true if atleast one element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 5];
assert!(arr.any_satisfy(|x| x % 2 == 1));

fn none_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: FnMut(&Self::Element) -> bool,

Returns true if none of elements in self satisfy pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [2, 4, 6];
assert!(arr.none_satisfy(|x| x % 2 == 1));

fn count_where<Pred>(&self, pred: Pred) -> usize
where Pred: FnMut(&Self::Element) -> bool,

Returns number of elements in self satisfying pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let n = arr.count_where(|x| x % 2 == 1);
assert_eq!(n, 2);

fn count_of(&self, e: &Self::Element) -> usize
where Self::Element: Eq,

Returns number of elements in self equals e.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [3, 0, 3];
let n = arr.count_of(&3);
assert_eq!(n, 2);

fn partition_point<F>(&self, belongs_in_second_half: F) -> Self::Position
where F: FnMut(&Self::Element) -> bool,

Returns position of first element of collection for which predicate returns false.

Precondition

The collection should be already partitioned wrt predicate i.e, there exist a position i such that predicate is true for every element of self.prefix_upto(i) and predicate is false for every element of self.suffix_from(i).

Complexity

O(log n) for RandomAccessCollection, O(n) otherwise; where n == self.count().

O(log n) application of belongs_in_second_half.

Example

use stl::*;

let arr = [1, 3, 5, 2, 4];
let i = arr.partition_point(|x| x % 2 == 0);
assert_eq!(i, 3);

fn fold_left<R, F>(&self, init: R, op: F) -> R
where F: FnMut(R, &Self::Element) -> R,

Returns the result of combining elements of given collection using given accumulation operation from left to right.

Postcondition

Result is (((init + e1) + e2) + ... + en). where e1, e2, …, en are the references to collection elements.

Complexity:

O(count)

Examples

use stl::*;

let arr = [1, 2, 3];
assert_eq!(arr.fold_left(0, |x, y| x + y), 6);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.