stl

Trait CollectionExt

pub trait CollectionExt: Collection {

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn first(&self) -> Option<Self::ElementRef<'_>> { ... }
    fn full(&self) -> Slice<'_, Self::Whole> { ... }
    fn prefix(&self, max_length: usize) -> Slice<'_, Self::Whole> { ... }
    fn prefix_upto(&self, pos: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn prefix_through(&self, pos: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn prefix_while<F: FnMut(&Self::Element) -> bool>(
        &self,
        predicate: F,
    ) -> Slice<'_, Self::Whole> { ... }
    fn dropping_while<F>(&self, predicate: F) -> Slice<'_, Self::Whole>
       where F: FnMut(&Self::Element) -> bool { ... }
    fn dropping_prefix(&self, count: usize) -> Slice<'_, Self::Whole> { ... }
    fn dropping_suffix(&self, count: usize) -> Slice<'_, Self::Whole> { ... }
    fn suffix(&self, max_length: usize) -> Slice<'_, Self::Whole> { ... }
    fn suffix_from(&self, from: Self::Position) -> Slice<'_, Self::Whole> { ... }
    fn splitting_at(
        &self,
        position: Self::Position,
    ) -> (Slice<'_, Self::Whole>, Slice<'_, Self::Whole>) { ... }
    fn splitting_after(
        &self,
        position: Self::Position,
    ) -> (Slice<'_, Self::Whole>, Slice<'_, Self::Whole>) { ... }
    fn iter(&self) -> CollectionIter<'_, Self::Whole>  { ... }
    fn for_each<F>(&self, f: F)
       where F: FnMut(&Self::Element) { ... }
    fn splitting_where<Pred>(
        &self,
        pred: Pred,
    ) -> SplitWhereIterator<'_, Self::Whole, Pred> 
       where Pred: FnMut(&Self::Element) -> bool,
             Self: Sized { ... }
    fn splitting_evenly_in_with_min_size(
        &self,
        max_slices: usize,
        min_size: usize,
    ) -> SplitEvenlyIterator<'_, Self::Whole>  { ... }
    fn splitting_evenly_in(
        &self,
        num_slices: usize,
    ) -> SplitEvenlyIterator<'_, Self::Whole>  { ... }
    fn map<MapFn, MappedType>(
        self,
        map_fn: MapFn,
    ) -> MappedCollection<Self, MapFn, MappedType>
       where Self: Sized,
             MapFn: Fn(&Self::Element) -> MappedType { ... }
    fn equals_by<OtherCollection, F>(
        &self,
        other: &OtherCollection,
        bi_pred: F,
    ) -> bool
       where OtherCollection: Collection,
             F: FnMut(&Self::Element, &OtherCollection::Element) -> bool { ... }
    fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
       where OtherCollection: Collection<Element = Self::Element>,
             Self::Element: Eq { ... }
    fn first_position_where<Pred>(&self, pred: Pred) -> Option<Self::Position>
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn first_position_of(&self, e: &Self::Element) -> Option<Self::Position>
       where Self::Element: Eq { ... }
    fn last_position_where<Pred>(&self, pred: Pred) -> Option<Self::Position>
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn last_position_of(&self, e: &Self::Element) -> Option<Self::Position>
       where Self::Element: Eq { ... }
    fn all_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn any_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn none_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn count_where<Pred>(&self, pred: Pred) -> usize
       where Pred: FnMut(&Self::Element) -> bool { ... }
    fn count_of(&self, e: &Self::Element) -> usize
       where Self::Element: Eq { ... }
    fn partition_point<F>(&self, belongs_in_second_half: F) -> Self::Position
       where F: FnMut(&Self::Element) -> bool { ... }
    fn partitioned<F>(
        &self,
        belongs_in_second_half: F,
    ) -> (Vec<Self::Element>, Vec<Self::Element>)
       where Self::Element: Clone,
             F: FnMut(&Self::Element) -> bool { ... }
    fn fold_left<R, F>(&self, init: R, op: F) -> R
       where F: FnMut(R, &Self::Element) -> R { ... }
    fn to_vec(&self) -> Vec<Self::Element>
       where Self::Element: Clone { ... }

}

Algorithms for Collection.

Provided Methods

fn is_empty(&self) -> bool

Returns true if given collection is empty.

Complexity

O(1).

fn first(&self) -> Option<Self::ElementRef<'_>>

Returns the first element, or nil if self is empty.

fn full(&self) -> Slice<'_, Self::Whole>

Returns slice of the collection covering full collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.full();
assert!(s.equals(&[1, 2, 3, 4, 5]));

fn prefix(&self, max_length: usize) -> Slice<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the initial elements of collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(k) where k is the number of elements in resultant slice.

Example

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.prefix(3);
assert!(s.equals(&[1, 2, 3]));

fn prefix_upto(&self, pos: Self::Position) -> Slice<'_, Self::Whole>

Returns prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let p = arr.prefix_upto(3);
assert!(p.equals(&[1, 2, 3]));

fn prefix_through(&self, pos: Self::Position) -> Slice<'_, Self::Whole>

Returns prefix slice of the collection ending at to inclusive.

Precondition

to is a valid position in the collection and to != self.end().

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let p = arr.prefix_through(3);
assert!(p.equals(&[1, 2, 3, 4]));

fn prefix_while<F: FnMut(&Self::Element) -> bool>( &self, predicate: F, ) -> Slice<'_, Self::Whole>

Returns prefix slice containing the initial elements until predicate returns false and skipping the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let arr = [1, 3, 5, 2, 7];
let p = arr.prefix_while(|x| x % 2 == 1);
assert!(p.equals(&[1, 3, 5]));

fn dropping_while<F>(&self, predicate: F) -> Slice<'_, Self::Whole>
where F: FnMut(&Self::Element) -> bool,

Returns a slice by skipping initial elements while predicate returns true, returning the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let arr = [1, 3, 5, 2, 4, 7];
let s = arr.dropping_while(|x| x % 2 == 1);
assert!(s.equals(&[2, 4, 7]));

fn dropping_prefix(&self, count: usize) -> Slice<'_, Self::Whole>

Returns a slice containing all but the given number of initial elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(k) otherwise, where k is number of elements to drop from beginning of collection.

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.dropping_prefix(3);
assert!(s.equals(&[4, 5]));

fn dropping_suffix(&self, count: usize) -> Slice<'_, Self::Whole>

Returns a slice containing all but the given number of final elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise, where n == self.count().

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.dropping_suffix(3);
assert!(s.equals(&[1, 2]));

fn suffix(&self, max_length: usize) -> Slice<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the final elements of the collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.suffix(3);
assert!(s.equals(&[3, 4, 5]));

fn suffix_from(&self, from: Self::Position) -> Slice<'_, Self::Whole>

Returns suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.suffix_from(3);
assert!(s.equals(&[4, 5]));

fn splitting_at( &self, position: Self::Position, ) -> (Slice<'_, Self::Whole>, Slice<'_, Self::Whole>)

Returns two disjoint slices of self split at the given position.

Examples

use stl::*;

let arr = [0, 1, 2, 3, 4];
let (s1, s2) = arr.splitting_at(2);
assert!(s1.equals(&[0, 1]));
assert!(s2.equals(&[2, 3, 4]));

fn splitting_after( &self, position: Self::Position, ) -> (Slice<'_, Self::Whole>, Slice<'_, Self::Whole>)

Returns two disjoint slices of self, split immediately after the given position.

Precondition

position != self.end().

Examples

use stl::*;

let arr = [0, 1, 2, 3, 4];
let (s1, s2) = arr.splitting_after(2);
assert!(s1.equals(&[0, 1, 2]));
assert!(s2.equals(&[3, 4]));

fn iter(&self) -> CollectionIter<'_, Self::Whole>

Returns an iterator to iterate over element refs in collection.

fn for_each<F>(&self, f: F)
where F: FnMut(&Self::Element),

Applies f to each element of collection.

Complexity:

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let mut sum = 0;
arr.for_each(|x| sum = sum + x);
assert_eq!(sum, 6);

fn splitting_where<Pred>( &self, pred: Pred, ) -> SplitWhereIterator<'_, Self::Whole, Pred>
where Pred: FnMut(&Self::Element) -> bool, Self: Sized,

Returns an iterator of slices which are separated by elements that match pred.

Example

use stl::*;

let arr = [1, 3, 5, 2, 2, 3, 4, 5, 5];

// Store sum of each split.
let mut res = vec![];
arr.splitting_where(|x| x % 2 == 0)
    .for_each(|s| res.push(s.iter().sum::<i32>()));
assert_eq!(res, vec![9, 0, 3, 10]);

fn splitting_evenly_in_with_min_size( &self, max_slices: usize, min_size: usize, ) -> SplitEvenlyIterator<'_, Self::Whole>

Returns an iterator that iterates through evenly sized consecutive at max max_slices slices of self with every slice being atleast of size min_size.

Precondition

max_slices > 0,

Postcondition

If splitting exactly evenly is not possible, then slices on start would have bigger size than slices at end, still maintaining as even splitting as possible.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr
    .splitting_evenly_in_with_min_size(3, 2)
    .map(|s| s.iter().copied().collect())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);

fn splitting_evenly_in( &self, num_slices: usize, ) -> SplitEvenlyIterator<'_, Self::Whole>

Returns an iterator that iterates through evenly sized consecutive num_slices slices of self.

Precondition

num_slices > 0.

Postcondition

If splitting exactly evenly is not possible, then slices on start would have bigger size than slices at end, still maintaining as even splitting as possible.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr
    .splitting_evenly_in(3)
    .map(|s| s.iter().copied().collect())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);

fn map<MapFn, MappedType>( self, map_fn: MapFn, ) -> MappedCollection<Self, MapFn, MappedType>
where Self: Sized, MapFn: Fn(&Self::Element) -> MappedType,

Returns a lazy collection projecting elements of mapping the given closure over elements.

Example

use stl::*;

let arr = [1, 2, 3, 4, 5].map(|x| x * 2);
assert!(arr.equals(&[2, 4, 6, 8, 10]));

fn equals_by<OtherCollection, F>( &self, other: &OtherCollection, bi_pred: F, ) -> bool
where OtherCollection: Collection, F: FnMut(&Self::Element, &OtherCollection::Element) -> bool,

Returns true if elements of self is equivalent to elements of other by given relation bi_pred.

Postcondition

Returns true if elements of self is equivalent to other by given relation bi_pred.

If self and other have different number of elements, then return false.

Complexity

O(min(m, n)) where

m == self.count()

n == other.count()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [2, 3, 4];
assert!(arr1.equals_by(&arr2, |x, y| *y == x + 1));

fn equals<OtherCollection>(&self, other: &OtherCollection) -> bool
where OtherCollection: Collection<Element = Self::Element>, Self::Element: Eq,

Returns true if elements of self is equal to elements of other.

Postcondition

Returns true if elements of self is equal to elements of other.

If self and other have different number of elements, then return false.

Complexity

O(min(m, n)) where

m == self.count()

n == other.count()

Examples

use stl::*;

let arr1 = [1, 2, 3];
let arr2 = [1, 2, 3];
assert!(arr1.equals(&arr2));

fn first_position_where<Pred>(&self, pred: Pred) -> Option<Self::Position>
where Pred: FnMut(&Self::Element) -> bool,

Finds position of first element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let i = arr.first_position_where(|x| *x == 3);
assert_eq!(i, Some(2));

fn first_position_of(&self, e: &Self::Element) -> Option<Self::Position>
where Self::Element: Eq,

Finds position of first element in self equals e. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 3];
let i = arr.first_position_of(&3);
assert_eq!(i, Some(2));

fn last_position_where<Pred>(&self, pred: Pred) -> Option<Self::Position>
where Pred: FnMut(&Self::Element) -> bool,

Finds position of last element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4];
let i = arr.last_position_where(|x| x % 2 == 1);
assert_eq!(i, Some(2));

fn last_position_of(&self, e: &Self::Element) -> Option<Self::Position>
where Self::Element: Eq,

Finds position of last element equals e. If no such element exist, return self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 3];
let i = arr.last_position_of(&3);
assert_eq!(i, Some(2));

fn all_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: FnMut(&Self::Element) -> bool,

Returns true if all element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 5];
assert!(arr.all_satisfy(|x| x % 2 == 1));

fn any_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: FnMut(&Self::Element) -> bool,

Returns true if atleast one element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 5];
assert!(arr.any_satisfy(|x| x % 2 == 1));

fn none_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: FnMut(&Self::Element) -> bool,

Returns true if none of elements in self satisfy pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [2, 4, 6];
assert!(arr.none_satisfy(|x| x % 2 == 1));

fn count_where<Pred>(&self, pred: Pred) -> usize
where Pred: FnMut(&Self::Element) -> bool,

Returns number of elements in self satisfying pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let n = arr.count_where(|x| x % 2 == 1);
assert_eq!(n, 2);

fn count_of(&self, e: &Self::Element) -> usize
where Self::Element: Eq,

Returns number of elements in self equals e.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [3, 0, 3];
let n = arr.count_of(&3);
assert_eq!(n, 2);

fn partition_point<F>(&self, belongs_in_second_half: F) -> Self::Position
where F: FnMut(&Self::Element) -> bool,

Returns position of first element of collection for which predicate returns false.

Precondition

The collection should be already partitioned wrt predicate i.e, there exist a position i such that predicate is true for every element of self.prefix_upto(i) and predicate is false for every element of self.suffix_from(i).

Complexity

O(log n) for RandomAccessCollection, O(n) otherwise; where n == self.count().

O(log n) application of belongs_in_second_half.

Example

use stl::*;

let arr = [1, 3, 5, 2, 4];
let i = arr.partition_point(|x| x % 2 == 0);
assert_eq!(i, 3);

fn partitioned<F>( &self, belongs_in_second_half: F, ) -> (Vec<Self::Element>, Vec<Self::Element>)
where Self::Element: Clone, F: FnMut(&Self::Element) -> bool,

Returns two Vec containing the elements of the collection that don’t and do satisfy the given predicate, respectively.

Postcondition

Returns (falseVec, trueVec) where falseVec contains all elements that don’t satisfy predicate and trueVec contains all elements that do satisfy predicate.

Relative ordering of elements is preserved in both Vec.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let (evens, odds) = arr.partitioned(|x| x % 2 == 1);
assert_eq!(evens, [2, 4]);
assert_eq!(odds, [1, 3, 5]);

fn fold_left<R, F>(&self, init: R, op: F) -> R
where F: FnMut(R, &Self::Element) -> R,

Returns the result of combining elements of given collection using given accumulation operation from left to right.

Postcondition

Result is (((init + e1) + e2) + ... + en). where e1, e2, …, en are the references to collection elements, where (a + b) represents op(a, b).

Complexity:

O(count)

Examples

use stl::*;

let arr = [1, 2, 3];
assert_eq!(arr.fold_left(0, |x, y| x + y), 6);

fn to_vec(&self) -> Vec<Self::Element>
where Self::Element: Clone,

Copies and returns all elements of self into Vec<Self::Element>.

Complexity

O(self.count())

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.