stl
Trait InputRange

pub trait InputRange {
    type Element;
    type Position: SemiRegular;

    // Required methods
    fn start(&self) -> Self::Position;
    fn end(&self) -> Self::Position;
    fn after(&self, i: Self::Position) -> Self::Position;
    fn at(&self, i: &Self::Position) -> &Self::Element;
}

Models a single-pass range.

Required Associated Types

type Element

Type of the element contained in self

type Position: SemiRegular

Type of the positions in self

Required Methods

fn start(&self) -> Self::Position

Returns the position of first element in self, or if self is empty then is_last_position(first_position()) == true

fn end(&self) -> Self::Position

Returns the “past the end” position in self, that is, the position immediately after the last element in self

fn after(&self, i: Self::Position) -> Self::Position

Returns position immediately after i

Requires: is_last_position(i) == false

fn at(&self, i: &Self::Position) -> &Self::Element

Access element at position i

Requires: i is a valid position in self and is_last_position(i) == false