stl

Trait LazyCollection

pub trait LazyCollection: Collection
where
    Self::Whole: LazyCollection,
{
    // Required method
    fn compute_at(&self, i: &Self::Position) -> Self::Element;
}

Models a collection whose elements are computed on element access.

Required Methods

fn compute_at(&self, i: &Self::Position) -> Self::Element

Computes element at position i.

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.