stl
Trait OutputRange

pub trait OutputRange: SemiOutputRange {
    type ElementMutRef<'a>: DerefMut<Target = Self::Element>
       where Self: 'a;

    // Required method
    fn at_mut<'a>(&'a mut self, i: &Self::Position) -> Self::ElementMutRef<'a>;
}

Models a fully mutable range.

OutputRange is extension of SemiOutputRange.

There are 2 types of safe mutation possible for ranges:

Reordering of elements that is provided by SemiOutputRange.

Element Assignment: Assign different value to element at any position. For the same, at_mut function is required that returns mutable reference to element at any position.

Required Associated Types

type ElementMutRef<'a>: DerefMut<Target = Self::Element> where Self: 'a

Type of mutable reference to element.

Very useful in scenarios when range actually doesn’t contain the elements and proxy mutable reference is necessary.

Required Methods

fn at_mut<'a>(&'a mut self, i: &Self::Position) -> Self::ElementMutRef<'a>

Access element at position i

Precondition

i is a valid position in self and i != end()

Complexity Requirement

O(1)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.