stl

Trait ParallelCollectionExt

pub trait ParallelCollectionExt: Collection
where
    Self::Whole: Send,
{
    // Provided methods
    fn parallel_first_position_where<Pred>(
        &self,
        pred: Pred,
    ) -> Option<Self::Position>
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn parallel_first_position_of(
        &self,
        e: &Self::Element,
    ) -> Option<Self::Position>
       where Self::Element: Eq + Sync { ... }
    fn parallel_last_position_where<Pred>(
        &self,
        pred: Pred,
    ) -> Option<Self::Position>
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn parallel_last_position_of(
        &self,
        e: &Self::Element,
    ) -> Option<Self::Position>
       where Self::Element: Eq + Sync { ... }
    fn parallel_all_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn parallel_any_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
    fn parallel_none_satisfy<Pred>(&self, pred: Pred) -> bool
       where Pred: Fn(&Self::Element) -> bool + Clone + Send { ... }
}

Parallel Algorithms for Collection.

Provided Methods

fn parallel_first_position_where<Pred>( &self, pred: Pred, ) -> Option<Self::Position>
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Finds position of first element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3];
let i = arr.parallel_first_position_where(|x| *x == 3);
assert_eq!(i, Some(2));

fn parallel_first_position_of( &self, e: &Self::Element, ) -> Option<Self::Position>
where Self::Element: Eq + Sync,

Finds position of first element in self equals e. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 3];
let i = arr.parallel_first_position_of(&3);
assert_eq!(i, Some(2));

fn parallel_last_position_where<Pred>( &self, pred: Pred, ) -> Option<Self::Position>
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Finds position of last element in self satisfying pred. If no such element exists, returns self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 3, 4];
let i = arr.parallel_last_position_where(|x| x % 2 == 1);
assert_eq!(i, Some(2));

fn parallel_last_position_of(&self, e: &Self::Element) -> Option<Self::Position>
where Self::Element: Eq + Sync,

Finds position of last element equals e. If no such element exist, return self.end().

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 3];
let i = arr.parallel_last_position_of(&3);
assert_eq!(i, Some(2));

fn parallel_all_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff all elements in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 3, 5];
assert!(arr.parallel_all_satisfy(|x| x % 2 == 1));

fn parallel_any_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff atleast one element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [1, 2, 5];
assert!(arr.any_satisfy(|x| x % 2 == 1));

fn parallel_none_satisfy<Pred>(&self, pred: Pred) -> bool
where Pred: Fn(&Self::Element) -> bool + Clone + Send,

Returns true iff no element in self satisfies pred.

Complexity

O(n) where n == self.count().

Example

use stl::*;

let arr = [2, 4, 6];
assert!(arr.none_satisfy(|x| x % 2 == 1));

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.