stl

Trait RandomAccessCollection

pub trait RandomAccessCollection: BidirectionalCollection<Position: Regular + Ord>
where
    Self::Whole: RandomAccessCollection,
{ }

Models a random access collection (similar to array) where jumping to any position from any other position is O(1) operation.

RandomAccessCollection is extension of BidirectionalCollection.

RandomAccessCollection enforces the Position of collection should be ordered.

RandomAccessCollection doesn’t add any new method but introduces complexity requirements mentioned below. The complexity requirements ensure that any position jump is O(1).

Complexity Requirements

rng.distance(from, to) -> O(1).

rng.form_next_n(i) -> O(1).

rng.form_prior_n(i) -> O(1).

rng.next_n(i) -> O(1).

rng.prior_n(i) -> O(1).

NOTE: If complexity requirements are not formed any algorithm on RandomAccessCollection have undefined behavior.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.