stl

Trait ReorderableCollectionExt

pub trait ReorderableCollectionExt: ReorderableCollection
where
    Self::Whole: ReorderableCollection,
{
    // Provided methods
    fn full_mut(&mut self) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_upto_mut(
        &mut self,
        to: Self::Position,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_through_mut(
        &mut self,
        pos: Self::Position,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_while_mut<F: FnMut(&Self::Element) -> bool>(
        &mut self,
        predicate: F,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn drop_while_mut<F>(&mut self, predicate: F) -> SliceMut<'_, Self::Whole>
       where F: FnMut(&Self::Element) -> bool { ... }
    fn drop_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn drop_end_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn suffix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn suffix_from_mut(
        &mut self,
        from: Self::Position,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn rotate(&mut self, at: Self::Position) -> Self::Position { ... }
}

Algorithms for ReorderableCollection.

Provided Methods

fn full_mut(&mut self) -> SliceMut<'_, Self::Whole>

Returns mutable slice of the collection covering full collection.

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut s = arr.full_mut();
let start = s.start();
*s.at_mut(&start) = 0;
assert!(s.equals(&[0, 2, 3, 4, 5]));
assert!(arr.equals(&[0, 2, 3, 4, 5]));

fn prefix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the initial elements of collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(k) where k is the number of elements in resultant slice.

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let s = arr.prefix_mut(3);
assert!(s.equals(&[1, 2, 3]));

fn prefix_upto_mut(&mut self, to: Self::Position) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut p = arr.prefix_upto_mut(3);
let start = p.start();
*p.at_mut(&start) = 0;
assert!(p.equals(&[0, 2, 3]));
assert!(arr.equals(&[0, 2, 3, 4, 5]));

fn prefix_through_mut( &mut self, pos: Self::Position, ) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice of the collection ending at to inclusive.

Precondition

to is a valid position in the collection and to != self.end().

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let p = arr.prefix_through_mut(3);
assert!(p.equals(&[1, 2, 3, 4]));

fn prefix_while_mut<F: FnMut(&Self::Element) -> bool>( &mut self, predicate: F, ) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice containing the initial elements until predicate returns false skipping the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let mut arr = [1, 3, 5, 2, 7];
let p = arr.prefix_while_mut(|x| x % 2 == 1);
assert!(p.equals(&[1, 3, 5]));

fn drop_while_mut<F>(&mut self, predicate: F) -> SliceMut<'_, Self::Whole>
where F: FnMut(&Self::Element) -> bool,

Returns a slice by skipping initial elements while predicate returns true, returning the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let mut arr = [1, 3, 5, 2, 4, 7];
let s = arr.drop_while_mut(|x| x % 2 == 1);
assert!(s.equals(&[2, 4, 7]));

fn drop_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole>

Returns a slice containing all but the given number of initial elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(k) otherwise, where k is number of elements to drop from beginning of collection.

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.drop(3);
assert!(s.equals(&[4, 5]));

fn drop_end_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole>

Returns a slice containing all but the given number of final elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise, where n == self.count().

Examples

use stl::*;

let arr = [1, 2, 3, 4, 5];
let s = arr.drop_end(3);
assert!(s.equals(&[1, 2]));

fn suffix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the final elements of the collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let s = arr.suffix_mut(3);
assert!(s.equals(&[3, 4, 5]));

fn suffix_from_mut(&mut self, from: Self::Position) -> SliceMut<'_, Self::Whole>

Returns mutable suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut s = arr.suffix_from_mut(3);
let start = s.start();
*s.at_mut(&start) = 0;
assert!(s.equals(&[0, 5]));
assert!(arr.equals(&[1, 2, 3, 0, 5]));

fn rotate(&mut self, at: Self::Position) -> Self::Position

Swaps the order in which the values self.prefix_upto(at) and self.suffix_from(at) occur, so that the end of the latter sequence of element values is the start of the former sequence, and returns that position.

Precondition

at is a valid position in self.

Complexity

O(n). At most n swaps. Where n == self.count().

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let i = arr.rotate(2);
assert_eq!(i, 3);
assert!(arr.equals(&[3, 4, 5, 1, 2]));

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.