stl

Trait ReorderableCollectionExt

pub trait ReorderableCollectionExt: ReorderableCollection
where
    Self::Whole: ReorderableCollection,
{

    // Provided methods
    fn full_mut(&mut self) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_upto_mut(
        &mut self,
        to: Self::Position,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_through_mut(
        &mut self,
        pos: Self::Position,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn prefix_while_mut<F: FnMut(&Self::Element) -> bool>(
        &mut self,
        predicate: F,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn dropping_while_mut<F>(
        &mut self,
        predicate: F,
    ) -> SliceMut<'_, Self::Whole>
       where F: FnMut(&Self::Element) -> bool { ... }
    fn dropping_prefix_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn dropping_suffix_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn suffix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole> { ... }
    fn suffix_from_mut(
        &mut self,
        from: Self::Position,
    ) -> SliceMut<'_, Self::Whole> { ... }
    fn splitting_at_mut(
        &mut self,
        position: Self::Position,
    ) -> (SliceMut<'_, Self::Whole>, SliceMut<'_, Self::Whole>) { ... }
    fn splitting_after_mut(
        &mut self,
        position: Self::Position,
    ) -> (SliceMut<'_, Self::Whole>, SliceMut<'_, Self::Whole>) { ... }
    fn splitting_where_mut<Pred>(
        &mut self,
        pred: Pred,
    ) -> SplitWhereIteratorMut<'_, Self::Whole, Pred> 
       where Pred: FnMut(&Self::Element) -> bool,
             Self: Sized { ... }
    fn splitting_evenly_in_with_min_size_mut(
        &mut self,
        max_slices: usize,
        min_size: usize,
    ) -> SplitEvenlyIteratorMut<'_, Self::Whole>  { ... }
    fn splitting_evenly_in_mut(
        &mut self,
        num_slices: usize,
    ) -> SplitEvenlyIteratorMut<'_, Self::Whole>  { ... }
    fn rotate(&mut self, at: Self::Position) -> Self::Position { ... }
    fn partition<F>(&mut self, belongs_in_second_partition: F) -> Self::Position
       where F: FnMut(&Self::Element) -> bool + Clone { ... }
    fn stable_partition<F>(
        &mut self,
        belongs_in_second_partition: F,
    ) -> Self::Position
       where F: FnMut(&Self::Element) -> bool + Clone { ... }

}

Algorithms for ReorderableCollection.

Provided Methods

fn full_mut(&mut self) -> SliceMut<'_, Self::Whole>

Returns mutable slice of the collection covering full collection.

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut s = arr.full_mut();
let start = s.start();
*s.at_mut(&start) = 0;
assert!(s.equals(&[0, 2, 3, 4, 5]));
assert!(arr.equals(&[0, 2, 3, 4, 5]));

fn prefix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the initial elements of collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(k) where k is the number of elements in resultant slice.

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let s = arr.prefix_mut(3);
assert!(s.equals(&[1, 2, 3]));

fn prefix_upto_mut(&mut self, to: Self::Position) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice of the collection ending at to exclusive.

Precondition

to is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut p = arr.prefix_upto_mut(3);
let start = p.start();
*p.at_mut(&start) = 0;
assert!(p.equals(&[0, 2, 3]));
assert!(arr.equals(&[0, 2, 3, 4, 5]));

fn prefix_through_mut( &mut self, pos: Self::Position, ) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice of the collection ending at to inclusive.

Precondition

to is a valid position in the collection and to != self.end().

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let p = arr.prefix_through_mut(3);
assert!(p.equals(&[1, 2, 3, 4]));

fn prefix_while_mut<F: FnMut(&Self::Element) -> bool>( &mut self, predicate: F, ) -> SliceMut<'_, Self::Whole>

Returns mutable prefix slice containing the initial elements until predicate returns false skipping the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let mut arr = [1, 3, 5, 2, 7];
let p = arr.prefix_while_mut(|x| x % 2 == 1);
assert!(p.equals(&[1, 3, 5]));

fn dropping_while_mut<F>(&mut self, predicate: F) -> SliceMut<'_, Self::Whole>
where F: FnMut(&Self::Element) -> bool,

Returns a slice by skipping initial elements while predicate returns true, returning the remaining elements.

Complexity

O(n) where n == self.count().

Examples

use stl::*;

let mut arr = [1, 3, 5, 2, 4, 7];
let s = arr.dropping_while_mut(|x| x % 2 == 1);
assert!(s.equals(&[2, 4, 7]));

fn dropping_prefix_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole>

Returns a slice containing all but the given number of initial elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(k) otherwise, where k is number of elements to drop from beginning of collection.

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let s = arr.dropping_prefix_mut(3);
assert!(s.equals(&[4, 5]));

fn dropping_suffix_mut(&mut self, count: usize) -> SliceMut<'_, Self::Whole>

Returns a slice containing all but the given number of final elements.

Postcondition

If count > self.count(), returns an empty slice.

Complexity

O(1) for RandomAccessCollection;

O(n) otherwise, where n == self.count().

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let s = arr.dropping_suffix_mut(3);
assert!(s.equals(&[1, 2]));

fn suffix_mut(&mut self, max_length: usize) -> SliceMut<'_, Self::Whole>

Returns a slice, upto specified maximum length, containing the final elements of the collection.

Postcondition

If the maximum length exceeds the number of elements in the collection, the result contains all elements in the collection.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let s = arr.suffix_mut(3);
assert!(s.equals(&[3, 4, 5]));

fn suffix_from_mut(&mut self, from: Self::Position) -> SliceMut<'_, Self::Whole>

Returns mutable suffix slice of the collection starting from from inclusive.

Precondition

from is a valid position in the collection.

Complexity

O(1).

Examples

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let mut s = arr.suffix_from_mut(3);
let start = s.start();
*s.at_mut(&start) = 0;
assert!(s.equals(&[0, 5]));
assert!(arr.equals(&[1, 2, 3, 0, 5]));

fn splitting_at_mut( &mut self, position: Self::Position, ) -> (SliceMut<'_, Self::Whole>, SliceMut<'_, Self::Whole>)

Returns two disjoint mutable slices of self split at the given position.

Examples

use stl::*;

let mut arr = [0, 1, 2, 3, 4];
let (s1, s2) = arr.splitting_at_mut(2);
assert!(s1.equals(&[0, 1]));
assert!(s2.equals(&[2, 3, 4]));

fn splitting_after_mut( &mut self, position: Self::Position, ) -> (SliceMut<'_, Self::Whole>, SliceMut<'_, Self::Whole>)

Returns two disjoint mutable slices of self, split immediately after the given position.

Precondition

position != self.end().

Examples

use stl::*;

let mut arr = [0, 1, 2, 3, 4];
let (s1, s2) = arr.splitting_after_mut(2);
assert!(s1.equals(&[0, 1, 2]));
assert!(s2.equals(&[3, 4]));

fn splitting_where_mut<Pred>( &mut self, pred: Pred, ) -> SplitWhereIteratorMut<'_, Self::Whole, Pred>
where Pred: FnMut(&Self::Element) -> bool, Self: Sized,

Returns an iterator of mutable slices which are separated by elements that match pred.

Example

use stl::*;

let mut arr = [1, 3, 5, 2, 2, 3, 4, 5, 7];
// Reverse each split
arr.splitting_where_mut(|x| x % 2 == 0).for_each(|mut s| s.reverse());
assert_eq!(arr, [5, 3, 1, 2, 2, 3, 4, 7, 5]);

fn splitting_evenly_in_with_min_size_mut( &mut self, max_slices: usize, min_size: usize, ) -> SplitEvenlyIteratorMut<'_, Self::Whole>

Returns an iterator that iterates through evenly sized consecutive at max max_slices mutable slices of self with every slice being atleast of size min_size.

Precondition

max_slices > 0,

Postcondition

If splitting exactly evenly is not possible, then slices on start would have bigger size than slices at end, still maintaining as even splitting as possible.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr
    .splitting_evenly_in_with_min_size_mut(3, 2)
    .map(|s| s.iter().copied().collect())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);

fn splitting_evenly_in_mut( &mut self, num_slices: usize, ) -> SplitEvenlyIteratorMut<'_, Self::Whole>

Returns an iterator that iterates through evenly sized consecutive num_slices mutable slices of self.

Precondition

num_slices > 0.

Postcondition

If splitting exactly evenly is not possible, then slices on start would have bigger size than slices at end, still maintaining as even splitting as possible.

Complexity

O(1) for RandomAccessCollection; otherwise O(n) where n == self.count().

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5, 6, 7];
let splits: Vec<Vec<_>> = arr
    .splitting_evenly_in_mut(3)
    .map(|s| s.iter().copied().collect())
    .collect();
assert_eq!(splits, vec![vec![1, 2, 3], vec![4, 5], vec![6, 7]]);

fn rotate(&mut self, at: Self::Position) -> Self::Position

Swaps the order in which the values self.prefix_upto(at) and self.suffix_from(at) occur, so that the end of the latter sequence of element values is the start of the former sequence, and returns that position.

Precondition

at is a valid position in self.

Complexity

O(n). At most n swaps. Where n == self.count().

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let i = arr.rotate(2);
assert_eq!(i, 3);
assert!(arr.equals(&[3, 4, 5, 1, 2]));

fn partition<F>(&mut self, belongs_in_second_partition: F) -> Self::Position
where F: FnMut(&Self::Element) -> bool + Clone,

Moves all elements satisfying the given predicate into a suffix of the collection, returning the start position of the resulting suffix.

Postcondition

If no elements exist in suffix, returns self.end().

Relative ordering of elements in the partitions are not preserved.

Complexity

O(count)

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let i = arr.partition(|x| x % 2 == 1);
assert_eq!(i, 2);
assert!(arr.prefix_upto(i).all_satisfy(|x| x % 2 == 0));
assert!(arr.suffix_from(i).all_satisfy(|x| x % 2 == 1));

fn stable_partition<F>( &mut self, belongs_in_second_partition: F, ) -> Self::Position
where F: FnMut(&Self::Element) -> bool + Clone,

Moves all elements satisfying the given predicate into a suffix of the given range, preserving the relative order of the elements in both partitions, and returns the start of the resulting suffix.

Postcondition

If no element exists in suffix, returns self.end().

Complexity

O(n log(n)) where n == self.count()

Example

use stl::*;

let mut arr = [1, 2, 3, 4, 5];
let i = arr.stable_partition(|x| x % 2 == 1);
assert_eq!(i, 2);
assert!(arr.equals(&[2, 4, 1, 3, 5]));

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.